% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-bbs.R
\name{auto_bbs}
\alias{auto_bbs}
\title{Automatically run a BBS analysis for a given species}
\usage{
auto_bbs(species = NULL, model = NULL, bbs_data = NULL,
  stratify_by = "bbs", output_dir = NULL, ...)
}
\arguments{
\item{species}{Character strings or vector of character strings of what
species are wanting to be analysed}

\item{model}{Character strings or vector of character strings of what
species are wanting to be analysed.}

\item{bbs_data}{Large list of BBS data as downloaded using \code{fetchBBSdata()}.
If this argument is left blank, \code{autoBBS()} will default to downloading
the latest data from the USGS website.}

\item{stratify_by}{Character string of how to stratify the BBS data}

\item{output_dir}{Path to directory where output should be directed to.
Defaults to current directory.}

\item{...}{Optional arguments for JAGS data prepping, modelling, or plotting.
See \code{run_model}, \code{prepare_jags_data}, and \code{generate_trend_plots}
for other argument details}
}
\value{
Some sort of output with cool stuff. Don't know yet
}
\description{
\code{auto_bbs} generates MCMC chains for a list of species using BBS
  data.
}
