% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoBBS.R
\name{autoBBS}
\alias{autoBBS}
\title{Automatically run a BBS analysis for a given species}
\usage{
autoBBS(species = NULL, model = NULL, outputDir = NULL, inits = NULL,
  params = c("n"), adaptSteps = 500, burnInSteps = 20000, nChains = 3,
  numSavedSteps = 2000, thinSteps = 10, nIter = ceiling((numSavedSteps *
  thinSteps)/nChains), runParallelChains = FALSE)
}
\arguments{
\item{species}{Character strings or vector of character strings of what
species are wanting to be analysed}

\item{model}{Character strings or vector of character strings of what
species are wanting to be analysed.}

\item{outputDir}{Path to directory where output should be directed to.
Defaults to current directory.}

\item{inits}{Optional list of initialization values for JAGS model.
If none are specified, the JAGS model will generate its own
initial values.}

\item{adaptSteps}{Optional integer specifying the number of steps to
adapt the JAGS model. It is recommended to do a minimum of 100,
\code{autoBBS} will default to 500.}

\item{burnInSteps}{Optional integer specifying the number of iterations
to burn in the model. Defaults to 20000 per chain.}

\item{nChains}{Optional number of chains to run. Defaults to 3.}

\item{numSavedSteps}{Optional number of steps to save per chain.
Defaults to 2000.}

\item{thinSteps}{Optional number of steps to thin or discard.}

\item{nIter}{Optional number of iterations.}

\item{runParallelChains}{Optional boolean value to specify whether
the chains are run in parallel. By default, the chains are not
run in parallel. If TRUE, the number of cores used will be the
minimum of number of chains and number of available cores.
JAGS console output is supressed with run in parallel.}
}
\value{
The sum of \code{x} and \code{y}
}
\description{
\code{autoBBS} generates MCMC chains for a list of species using BBS
  data.
}
\examples{
add(1, 1)
add(10, 1)
}
