% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_data_sample}
\alias{bbs_data_sample}
\title{Sample BBS data}
\format{
An object of class \code{list} of length 3.
}
\usage{
bbs_data_sample
}
\description{
Contains only Pacific Wren data
}
\keyword{datasets}
