% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_models}
\alias{bbs_models}
\title{List of models available}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 3 columns.
}
\usage{
bbs_models
}
\description{
List of models available
}
\keyword{datasets}
