% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_strata}
\alias{bbs_strata}
\title{List of established strata}
\format{
An object of class \code{list} of length 5.
}
\usage{
bbs_strata
}
\description{
Each list item contains a data frame describing the strata for that
stratification.
}
\details{
Contains \code{bbs_usgs}, \code{bbs_cws}, \code{bcr}, \code{latlong} and \code{prov_state}
}
\keyword{datasets}
