% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsBayes-defunct.R
\name{bbsBayes-defunct}
\alias{bbsBayes-defunct}
\alias{generate_cont_indices}
\alias{generate_regional_indices}
\alias{generate_strata_indices}
\alias{generate_cont_trend}
\alias{generate_regional_trends}
\alias{generate_strata_trends}
\alias{plot_cont_indices}
\alias{prepare_jags_data}
\alias{p_waic}
\alias{waic}
\title{bbsBayes defunct functions}
\usage{
generate_cont_indices()

generate_regional_indices()

generate_strata_indices()

generate_cont_trend()

generate_regional_trends()

generate_strata_trends()

plot_cont_indices()

prepare_jags_data()

p_waic()

waic()
}
\description{
\code{generate_indices()} supersedes
\itemize{
\item \code{generate_cont_indices()}
\item \code{generate_regional_indices()}
\item \code{generate_strata_indices()}
}

\code{generate_trends()} supersedes
\itemize{
\item \code{generate_cont_trend()}
\item \code{generate_regional_trends()}
\item \code{generate_strata_trends()}
}

\code{plot_indices()} supersedes
\itemize{
\item \code{plot_cont_indices()}
}

\code{prepare_data()} supersedes
\itemize{
\item \code{prepare_jags_data()}
}

\code{p_waic()} and \code{waic()} are no longer recommended for BBS data. Use
cross validation instead
}
