% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{birds}
\alias{birds}
\title{Bird Species Counts by Route}
\format{A data frame with 6189135 rows and 20 variables:
\describe{
 \item{statenum}{Numerical representation of political state.}
 \item{Route}{Breeding Bird Survey route number.}
 \item{countrynum}{Numerical representation of country}
 \item{RPID}{}
 \item{Year}{Year of the given count.}
 \item{Aou}{Numerical species code as assigned by American Ornithological
   Union}
 \item{count10}{}
 \item{count20}{}
 \item{count30}{}
 \item{count40}{}
 \item{count50}{}
 \item{StopTotal}{}
 \item{SpeciesTotal}{}
 \item{sp.bbas}{Numerical species code as assigned by BBS}
 \item{BCR}{Numerical representation of geopolitical stratum.}
 \item{route}{Breeding Bird Survey route number.}
 \item{state}{Numerical representation of political state.}
 \item{rt.uni}{Combined state-route}
 \item{rt.uni.y}{Combined state-route-year}
 \item{runyear}{Year of the route run.}
}}
\usage{
birds
}
\description{
A dataset containing counts of individual bird species
by route per year.
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
