% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{copy_model_file}
\alias{copy_model_file}
\title{Copy model file}
\usage{
copy_model_file(model, model_variant, dir, overwrite = FALSE)
}
\arguments{
\item{model}{Character. Type of model to use, must be one of "first_diff"
(First Differences), "gam" (General Additive Model), "gamye" (General
Additive Model with Year Effect), or "slope" (Slope model).}

\item{model_variant}{Character. Model variant to use, must be one of
"nonhier" (Non-hierarchical), "hier" (Hierarchical; default), or "spatial"
(Spatially explicit).}

\item{dir}{Character. Directory where file should be saved.}

\item{overwrite}{Logical. Whether to overwrite an existing copy of the model
file.}
}
\value{
File path to copy of the model file.
}
\description{
Save a predefined Stan model file to a local text file for editing. These
files can then be used in \code{run_models()} by specifying the \code{model_file}
argument.
}
\examples{

# Save the Slope model in temp directory
copy_model_file(model = "slope", model_variant = "spatial", dir = tempdir())

# Overwrite an existing copy
copy_model_file(model = "slope", model_variant = "spatial", dir = tempdir(),
                overwrite = TRUE)

# Clean up
unlink(file.path(tempdir(), "slope_spatial_bbs_CV_COPY.stan"))

}
