% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datacount.sp}
\alias{datacount.sp}
\title{Count Summaries by Species}
\format{A data frame with 663 rows and 7 variables:
\describe{
 \item{sp}{Numerical representation of species assigned by BBS.}
 \item{eng}{English name of the bird species.}
 \item{nprov.all}{}
 \item{nroute.all}{}
 \item{nyear.all}{Number of years that this species has been counted.}
 \item{nrouteyears.all}{}
 \item{rungroup}{Numerical values from 1-5 that allow species to
   be run in batches in desired.}
}}
\usage{
datacount.sp
}
\description{
A dataset containing summary totals for each BBS species.
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
