% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-index-data.R
\name{extract_index_data}
\alias{extract_index_data}
\title{Extract index data}
\usage{
extract_index_data(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\description{
Subsets area weights and extracts n values from JAGS samples
}
\keyword{internal}
