% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-index-data.R
\name{extract_index_data_stan}
\alias{extract_index_data_stan}
\title{Extract index data (STAN)}
\usage{
extract_index_data_stan(model_fit = NULL, alt_n = "n", model_data = NULL)
}
\arguments{
\item{model_fit}{Model fit list generated by \code{run_model}}
}
\value{
List of index values produced by JAGS samples
}
\description{
Subsets area weights and extracts n values from JAGS samples
}
\keyword{internal}
