% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-index-data.R
\name{extract_index_data_tidy}
\alias{extract_index_data_tidy}
\title{Extract index data (TIDY)}
\usage{
extract_index_data_tidy(jags_mod = NULL, alt_n = "n", jags_data = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
List of index values produced by JAGS samples
}
\description{
Subsets area weights and extracts n values from JAGS samples
}
\keyword{internal}
