% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{fetch_bbs_data}
\alias{fetch_bbs_data}
\title{Fetch Breeding Bird Survey dataset}
\usage{
fetch_bbs_data()
}
\value{
Large list (3 elements) containing point count data, route data, and species data.
}
\description{
\code{fetch_bbs_data} uses File Transfer Protocol (FTP) to fetch Breeding Bird
 Survey data from the United States Geological Survey (USGS) FTP site. This
 is the raw data that is uploaded to the site before any analyses are performed.
}
\examples{
bbs_data <- fetch_bbs_data() # Be sure to save the output in an R object!

}
