% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-indices.R
\name{generate_cont_indices}
\alias{generate_cont_indices}
\title{Generate a dataframe of the continental yearly indices}
\usage{
generate_cont_indices(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Data frame of 4 variables:
  \item{Year}{Year of particular index}
  \item{Index}{Strata-weighted count index}
  \item{Q25}{2.5\% quantile of mean strata-weighted indices}
  \item{Q975}{97.5\% quantile of mean strata-weighted indices}
}
\description{
\code{generate_cont_indices} creates a data frame of the strata-weighed
  continental indices by year. This data frame can then be used to
  generate a population trajectory of the species.
}
\examples{

\dontrun{
# After running a JAGS model, use that output to create trend data frame
cont_index <- generate_cont_indices(jags_mod = mod)
}

}
