% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-trend.R
\name{generate_cont_trend}
\alias{generate_cont_trend}
\title{Generate the continental trend}
\usage{
generate_cont_trend(indices = NULL, min_year = NULL, max_year = NULL)
}
\arguments{
\item{indices}{Continental indices generated by \code{generate_cont_indices}}

\item{min_year}{Minimum year to calculate trends from}

\item{max_year}{Maximum year to calculate trends to}
}
\value{
Numeric percentage of trend
}
\description{
\code{generate_cont_trend} calculates the geometric mean annual changes in population size.
}
\examples{

\dontrun{
# After running a JAGS model, use that output to create trend data frame
cont_trend <- generate_cont_trend(indices = indices)
}

}
