% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map.R
\name{generate_map}
\alias{generate_map}
\title{Generate a map of trends by strata.}
\usage{
generate_map(trend = NULL, stratify_by = NULL)
}
\arguments{
\item{trend}{Dataframe of strata trends produced by
\code{generate_strata_trends}}

\item{stratify_by}{How was the data stratified?}
}
\value{
spplot object
}
\description{
\code{generate_map} allows you to generate a colour-coded map of species trends
  for each strata. Given trends generated by \code{generate_strata_trends}, this
  function will shade in each stratum based on the percent
  change in that stratum.
}
\examples{

\dontrun{
# Suppose we stratified our data using "bbs_usgs"
strata_map <- generate_map(trend = generate_strata_trends(indices),
                           stratify_by = "bbs_usgs")
}
}
