% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-regional-indices.R
\name{generate_regional_indices}
\alias{generate_regional_indices}
\title{Generate a dataframe of the regional yearly indices}
\usage{
generate_regional_indices(jags_mod = NULL, jags_data = NULL,
  quantiles = c(0.025, 0.05, 0.25, 0.75, 0.95, 0.975),
  regions = c("stratum", "continental"), alternate_n = "n",
  startyear = NULL, max_backcast = 5, alt_region_names = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{jags_data}{data object used in \code{run_model}}

\item{quantiles}{vector of quantiles to be sampled from the posterior distribution Defaults to c(0.025,0.05,0.25,0.5,0.75,0.95,0.975)}

\item{regions}{vector selcting regional compilation(s) to calculate. Default is "continental","stratum", options also include "national", "prov_state", "bcr", and "bcr_by_country" for the stratifications that include areas that align with those regions.}

\item{alternate_n}{text string indicating the name of the alternative annual index parameter in a model, Default is "n"}

\item{startyear}{Optional first year for which to calculate the annual indices if a trajectory for only the more recent portion of the time series is desired. This is probably most relevant if max_backcast is set and so trajectories for different time-periods could include a different subset of strata (i.e., strata removed)}

\item{max_backcast}{an integer indicating the maximum number of years to backcast the stratum-level estimates before the first year in which the species was observed on any route in that stratum. Default is 5. If the observed data in a given stratum do not include at least one non-zero observation of the species between the first year of the BBS and startyear+max_backcast, the stratum is dropped from the relevant regional summary. Optionally this can be set to NULL to ignore any backcasting limit (i.e., to generate annual indices for the entire time series, regardless of when the species was first observed)}

\item{alt_region_names}{Optional dataframe generated with the \code{extract_strata_areas} function, then modified with an additional column indicating the strata to include in a custom spatial summary}
}
\value{
List of 6 objects:
  \item{data_summary}{dataframe with the following columns}
  \item{Year}{Year of particular index}
  \item{Region}{Region name}
  \item{Region_alt}{Long name for region}
  \item{Region_type}{Type of region including "continental", "national","Province/State","BCR", "bcr_by_country", or "stratum"}
  \item{Strata_included}{Strata included in the annual index calculations}
  \item{Strata_excluded}{Strata excluded from the annual index calculations because they have no observations of the species in the first part of the time series, see arguments max_backcast and startyear}
  \item{Index}{Strata-weighted count index}
  \item{additional columns for each of the values in quantiles}{quantiles of the posterior distribution}
  \item{obs_mean}{Mean of the observed annual counts of birds across all routes and all years. An alternative estimate of the average relative abundance of the species in the region and year. Differences between this and the annual indices are a function of the model. For composite regions (i.e., anything other than stratum-level estimates) this average count is calculated as an area-weighted average across all strata included}
  \item{nrts}{Number of BBS routes that contributed data for this species, region, and year}
  \item{nnzero}{Number of BBS routes on which this species was observed (i.e., count is > 0) in this region and year}

  \item{samples}{array of all samples from the posterior distribution}
  \item{area-weights}{data frame of the strata names and area weights used to calculate the continental estimates}
  \item{y_min}{first year used in the summary, scale 1:length of time-series}
  \item{y_max}{last year used in the summary, scale 1:length of time-series}
  \item{startyear}{first year used in the summary, scale 1966:2018}
}
\description{
\code{generate_regional_indices} creates a data frame of the strata-weighted
  regional indices by year. This data frame can then be used to
  generate a population trajectory of the species.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_index <- generate_cont_indices(jags_mod = mod)
}

}
