% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-regional-trends.R
\name{generate_regional_trends}
\alias{generate_regional_trends}
\title{Generate regional trends for countries, states/provinces, or BCRs from analyses run on the cws or usgs stratifications}
\usage{
generate_regional_trends(indices = NULL, min_year = NULL,
  max_year = NULL, quantiles = c(0.025, 0.05, 0.25, 0.75, 0.95, 0.975),
  slope = FALSE)
}
\arguments{
\item{indices}{regional indices generated by \code{generate_regional_indices}}

\item{min_year}{Minimum year to calculate trends from}

\item{max_year}{Maximum year to calculate trends to}

\item{quantiles}{vector of quantiles to be sampled from the posterior distribution Defaults to c(0.025,0.05,0.25,0.5,0.75,0.95,0.975)}

\item{slope}{Logical, if TRUE, calculates an alternative trend metric, the slope of a log-linear regression through the annual indices. Default FALSE}
}
\value{
Numeric percentage of trend
}
\description{
\code{generate_regional_trends} calculates the geometric mean annual changes in population size for composite regions.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bbs_usgs")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the annual indices for states/provinces, countries, and BCRs
regional_index <- generate_regional_indices(jags_mod = mod)

#Output the trend
generate_regional_trends(indices = regional_index)
}

}
