% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-rmd-report.R
\name{generate_rmd_report}
\alias{generate_rmd_report}
\title{Generate R Markdown report}
\usage{
generate_rmd_report(jags_mod = NULL, output_dir = getwd(), ...)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{...}{Other arguments}
}
\value{
List of ggplot2 objects
}
\description{
Generates an R Markdown dashboard report that allows you to
  explore all trend plots generated as well as MCMC diagnostics.
}
\examples{

\dontrun{
}

}
