% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-strata-trend.R
\name{generate_strata_trend}
\alias{generate_strata_trend}
\title{Generate a dataframe of the strata trend}
\usage{
generate_strata_trend(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Data frame of 4 variables:
  \item{Year}{Year of particular index}
  \item{Index}{Strata-weighted count index}
  \item{Q25}{2.5\% quantile of mean strata-weighted indices}
  \item{Q975}{97.5\% quantile of mean strata-weighted indices}
  \item{Stratum}{Name of the stratum}
}
\description{
\code{generate_strata_trend} creates a dataframe of trend indices for
  each strata modelled. This is useful if you would like
  to plot the year-to-year trend using your own plot recipes.
}
