% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-trend.R
\name{generate_trend}
\alias{generate_trend}
\title{Generate trend indices for a given species.}
\usage{
generate_trend(jags_data = NULL, stratify_by = NULL)
}
\arguments{
\item{jags_data}{List generated by \code{run_model} containing the JAGS model
and MCMC samples}

\item{stratify_by}{How was the data stratified?}
}
\value{

}
\description{
Generates yearly indices for the specified bird species
  given simulated MCMC chains
}
\examples{

\dontrun{
}

}
