% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-trend-plot.R
\name{generate_trend_plot}
\alias{generate_trend_plot}
\title{Generate trend indices for a given species.}
\usage{
generate_trend_plot(jags_mod = NULL, strata_to_plot = NULL, cont = TRUE,
  ...)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{strata_to_plot}{Which strata would you like to include in the list
of plots? Provide a list of strata to plot or the string "all"
to plot all the strata used in the analysis.}

\item{cont}{Should a continental plot be produced?}

\item{...}{Other arguments}
}
\value{
List of ggplot2 objects
}
\description{
Generates yearly indices for the specified bird species
  given simulated MCMC chains.
}
\examples{

\dontrun{
# Given we've run all stratification and prep functions, run model
jags_mod <- run_model(jags_data = data_jags,
                      model = "standard")

# Generate continental plots only
plots <- generate_trend_plot(jags_mod = jags_mod)

# Generate continental trend plots and trend plots for all strata
plots <- generate_trend_plot(jags_mod = jags_mod,
                             strata_to_plot = "all")

# Generate continental trend plots and trend plots for specific strata
plots <- generate_trend_plot(jags_mod = jags_mod,
                             strata_to_plot = c("US-GA-27", "US-FL-31"))

# Don't generate a continental plot
plots <- generate_trend_plot(jags_mod = jags_mod,
                             strata_to_plot = "all",
                             cont = FALSE)


}

}
