% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-trends.R
\name{generate_trends}
\alias{generate_trends}
\title{Generate regional trends}
\usage{
generate_trends(
  indices,
  min_year = NULL,
  max_year = NULL,
  quantiles = c(0.025, 0.05, 0.25, 0.75, 0.95, 0.975),
  slope = FALSE,
  prob_decrease = NULL,
  prob_increase = NULL
)
}
\arguments{
\item{indices}{List. Indices generated by \code{generate_indices()}

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{quantiles}{Numeric vector. Quantiles to be sampled from the posterior
distribution. Defaults to \code{c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975)}}

\item{slope}{Logical. Whether to calculate an alternative trend metric, the
slope of a log-linear regression through the annual indices. Default
\code{FALSE}.}

\item{prob_decrease}{Numeric vector. Percent-decrease values for which to
optionally calculate the posterior probabilities (see Details). Default is
\code{NULL} (do not calculate).}

\item{prob_increase}{Numeric vector. Percent-increase values for which to
optionally calculate the posterior probabilities (see Details). Default is
\code{NULL} (do not calculate).}
}
\value{
Data frame with one row for each region included in the starting \code{indices} object. Contains \code{start_year}, \code{end_year}, \code{region}, \code{region-alt}, \code{region_type},
\code{strata_included}, \code{strata_excluded}
\item{Start_year}{first year of the trend}
\item{End_year}{last year of the trend}
\item{Region}{short name of the region}
\item{Region_alt}{Long name for region}
\item{Region_type}{Type of region including continental, national,Province_State,BCR, bcr_by_national, or stratum}
\item{Strata_included}{Strata included in the trend and annual index calculations}
\item{Strata_excluded}{Strata potentially excluded from the trend and annual index calculations because they have no observations of the species in the first part of the time series}
\item{Trend}{Estimated mean annual percent change over the trend time-period (i.e., Start_year - End_year), according to an endpoint comparison of annual index in Start_year and the annual index in End_year}
\item{Trend_Q_quantiles}{quantiles of the posterior distribution of Trend estimates, matching levels included in the quantiles argument}
\item{Percent_Change}{Estimated total percent change over the trend time-period}
\item{Percent_Change_Q_quantiles}{quantiles of the posterior distribution of Percent Change estimates, matching levels included in the quantiles argument}
\item{Slope_Trend}{Estimated mean annual percent change over the trend time-period, according to the slope of a linear regression through the log-transformed annual indices}
\item{Slope_Trend_Q_quantiles}{quantiles of the posterior distribution of Percent Change estimates, matching levels included in the quantiles argument}
\item{prob_decrease_X_percent}{proportion of the posterior distribution of Percent_Change that is below the percentage values supplied in prob_decrease}
\item{prob_increase_X_percent}{proportion of the posterior distribution of Percent_Change that is above the percentage values supplied in prob_increase}
\item{Relative_Abundance}{Mean of the annual index values across all years. An estimate of the average relative abundance of the species in the region. Can be interepreted as the predicted average count of the species in an average year on an average route by an average observer, for the years, routes, and observers in the existing data}
\item{Observed_Relative_Abundance}{Mean of the observed annual counts of birds across all routes and all years. An alternative estimate of the average relative abundance of the species in the region. For composite regions (i.e., anything other than stratum-level estimates) this average count is calculated as an area-weighted average across all strata included}
\item{Number_of_Strata}{The number of strata included in the region}
\item{Width_of_X_percent_Credible_Interval}{Width (in percent/year) of the credible interval on the Trend calculation. Calculated for the widest credible interval requested in quantiles argument. Default is 95 percent CI (i.e., Trend_Q0.975 - Trend_Q0.025)}
\item{Width_of_X_percent_Credible_Interval_Slope}{Width (in percent/year) of the credible interval on the Trend calculation for the slope-based trend. Calculated for the widest credible interval requested in quantiles argument. Default is 95 percent CI (i.e., Slope_Trend_Q0.975 - Slope_Trend_Q0.025)}
\item{Number_of_Routes}{The number of unique BBS routes included in the annual indices for this region and species, i.e., number of routes for this region and species for the years since \code{generate_indices(startyear)}}
\item{Mean_Number_of_Routes}{The average number of BBS routes across years contributing data for this region and species}
\item{backcast_flag}{approximate proportion of the included species range*years that are supported by data in a given region and year, e.g., 1.0 = data cover full time-series, 0.75 = data cover 75 percent of time-series. Only calculated if max_backcast != NULL}
}
\description{
Generates trends for continent and strata and optionally for countries,
states/provinces, or BCRs from analyses run on the stratifications that
support these composite regions. Calculates the geometric mean annual changes
in population size for composite regions.
}
\details{
The posterior probabilities can be calculated for a percent-decrease
(\code{prob_decrease}) and/or percent-increase (\code{prob_increase}) if desired.
These calculate the probability that the populationhas decreased/increased
by at least the amount specified.

For example, a \code{prob_increase = 100} would result in the calculation of the
probability that the population has increased by more than 100\% (i.e.,
doubled) over the period of the trend.

Alternatively, a \code{prob_decrease = 50} would result in the calculation of
the probability that the population has decreased by more than 50\% (i.e.,
less than half of the population remains) over the period of the trend.
}
\examples{

# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in a JAGS model.
d <- prepare_data(s, species = "Pacific Wren",
                  min_year = 2009,
                  max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)

# Generate the continental and stratum indices
i <- generate_indices(model_output = m)

# Now, generate the trends
t <- generate_trends(i)


}
