% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_convergence}
\alias{get_convergence}
\title{Convergence metrics}
\usage{
get_convergence(model_output, variables = NULL)
}
\arguments{
\item{model_output}{List. Output of \code{run_model()}.}
}
\value{
Data frame of convergence metrics for all model variables. Contains
\verb{variable_type, }variable\verb{, }ess_bulk\verb{, }ess_tail\verb{, and }rhat`
}
\description{
Convergence metrics
}
\examples{

# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in modelling
d <- prepare_data(s, min_year = 2009, max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)

# Calculate convergence metrics on each variable, a variable type, or on a
# single specific variable
conv <- get_convergence(m)
conv <- get_convergence(m, variables = "strata_raw")
conv <- get_convergence(m, variables = "strata_raw[9]")

}
