% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_model_vars}
\alias{get_model_vars}
\title{Get model variables}
\usage{
get_model_vars(model_output, all = FALSE)
}
\arguments{
\item{model_output}{List. Output of \code{run_model()}.}
}
\description{
Returns the basic model variables types (note that most variables have
different iterations for each strata and each year).
}
\examples{

# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in modelling
d <- prepare_data(s, min_year = 2009, max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)

# Calculate convergence metrics on each variable
get_model_vars(m)

}
