% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-species-aou.R
\name{get_species_aou}
\alias{get_species_aou}
\title{Get AOU code of a species}
\usage{
get_species_aou(species, sp_eng)
}
\arguments{
\item{species}{Species List}

\item{sp_eng}{English name of species}
}
\value{
4-5 digit AOU code for the given species, -1 if not found
}
\description{
Simply return the 4-5 digit AOU code for the given species.
}
\keyword{internal}
