% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{have_bbs_data}
\alias{have_bbs_data}
\title{Check whether BBS data exists}
\usage{
have_bbs_data(level = "state", release = 2022)
}
\value{
\code{TRUE} if the data is found, \code{FALSE} otherwise
}
\description{
Check whether BBS data exists
}
