% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppd.R
\name{lppd}
\alias{lppd}
\title{Calculate log posterior predictive density}
\usage{
lppd(jags_data = NULL, jags_mod = NULL, pointwise = FALSE)
}
\arguments{
\item{jags_data}{Data prepared by \code{prepare_jags_data}, used
for input to the JAGS model}

\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{pointwise}{If set to \code{TRUE}, a data frame is returned
that contains the pointwise LPPD for each count. Defaults
to \code{FALSE}}
}
\value{
Data frame of pointwise LPPD by count if \code{pointwise}
  is set to \code{TRUE}. Double precision numerical value of LPPD if
  \code{pointwise} is set to \code{FALSE}.
}
\description{
\code{lppd} Calculate log posterior preditive density (LPPD) for the supplied
  model.
}
\details{
NOTE: in order to calculated LPPD, the model MUST track the parameter
  "lambda". In species that are data-rich, such as Barn Swallow,
  this produces extremely large JAGS objects, and takes up a considerable
  amount of memory when simulating with \code{run_model}
}
\examples{
\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)


  # Output LPPD
  lppd(jags_data = prepped_data,
       jags_mod = mod)

  # Get dataframe of pointwise LPPD values for each count
  lppd_pointwise <- lppd(jags_data = prepped_data,
                         jags_mod = mod,
                         pointwise = TRUE)
}

}
