% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{midyear}
\alias{midyear}
\title{Midyear of BBS Counts}
\format{Integer:
\describe{
 \item{midyear}{Floor of the mid year between the first count year and
  current count year, i.e. floor((Current Year + 1966)/2)}
}}
\usage{
midyear
}
\description{
The middle year of BBS counts, the first count being in 1966 up to
this data set's year of 2016.
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
