% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p-waic.R
\name{p_waic}
\alias{p_waic}
\title{Calculate effective number of parameters}
\usage{
p_waic(jags_data = NULL, jags_mod = NULL, pointwise = FALSE)
}
\arguments{
\item{jags_data}{Data prepared by \code{prepare_jags_data}, used
for input to the JAGS model}

\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{pointwise}{If set to \code{TRUE}, a data frame is returned
that contains the pointwise LPPD for each count. Defaults
to \code{FALSE}}
}
\value{
Data frame of pointwise pWAIC by count if \code{pointwise}
  is set to \code{TRUE}. Double precision numerical value of pWAIC if
  \code{pointwise} is set to \code{FALSE}.
}
\description{
\code{p_waic} calculates effective number of parameters. This is
  used as a bias adjustment factor in the calculation of WAIC
  in the function \code{waic}.
}
\details{
NOTE: in order to calculated pWAIC, the model MUST track the parameter
  "lambda". In species that are data-rich, such as Barn Swallow,
  this produces extremely large JAGS objects, and takes up a considerable
  amount of memory when simulating with \code{run_model}

See examples for details.
}
\examples{
\dontrun{
  # Run a JAGS model, and be sure to track the parameter "lambda"
  jags_mod <- run_model(jags_data = prepared_data,
                        parameters_to_save = "lambda")

  # Output the pWAIC value
  p_waic(jags_data = prepared_data,
         jags_mod = jags_mod)
}

}
