% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacific_wren_model}
\alias{pacific_wren_model}
\title{Example model output}
\format{
An object of class \code{list} of length 4.
}
\usage{
pacific_wren_model
}
\description{
Pacific wren model output
}
\keyword{datasets}
