% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-continental.R
\name{plot_continental}
\alias{plot_continental}
\title{Generate continental trend plot.}
\usage{
plot_continental(n = NULL, area_weights = NULL, y_min = NULL,
  y_max = NULL, r_year = NULL)
}
\arguments{
\item{n}{Posterior samples of n}

\item{area_weights}{Area weights of strata that were used}

\item{y_min}{Minimum year}

\item{y_max}{Maximum year}

\item{r_year}{Actual years that were used}
}
\value{
ggplot2 object of continental trend plot.
}
\description{
Generates the trend plot for continental indices.
}
\keyword{internal}
