% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_map}
\alias{plot_map}
\title{Generate a map of trends by strata}
\usage{
plot_map(trends, slope = FALSE, title = TRUE, col_viridis = TRUE)
}
\arguments{
\item{trends}{Data frame. Trends produced by \code{generate_trends()}.}

\item{slope}{Logical. Whether or not to map values of the alternative trend
metric (slope of a log-linear regression) if \code{slope = TRUE} was used in
\code{generate_trends()},  through the annual indices. Default \code{FALSE}.}

\item{title}{Logical. Whether or not to include a title with species. Default
\code{TRUE}.}

\item{col_viridis}{Logical flag to use "viridis" colour-blind friendly
palette. Default is FALSE}
}
\value{
spplot object
}
\description{
\code{plot_map()} allows you to generate a colour-coded map of there percent
change in species trends for each strata.
}
\examples{
# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in a JAGS model.
d <- prepare_data(s, min_year = 2009, max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)

# Generate the continental and stratum indices
i <- generate_indices(m)

# Now, generate the trends
t <- generate_trends(i)

# Generate the map
map <- plot_map(t)

}
