% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-strata.R
\name{plot_strata}
\alias{plot_strata}
\title{Generate continental trend plot.}
\usage{
plot_strata(n = NULL, strata_to_plot = NULL, area_weights = NULL,
  y_min = NULL, y_max = NULL, r_year = NULL)
}
\arguments{
\item{n}{Posterior samples of n}

\item{strata_to_plot}{Which strata to plot}

\item{area_weights}{Subsetted area weights, used more for the names of strata}

\item{y_min}{Minimum year}

\item{y_max}{Maximum year}

\item{r_year}{Actual years that were used}
}
\value{
List of ggplots for each strata
}
\description{
Generates the trend plot for continental indices.
}
\keyword{internal}
