% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-strata-trend.R
\name{plot_strata_trend}
\alias{plot_strata_trend}
\title{Generate trend plot by stratum.}
\usage{
plot_strata_trend(trend = NULL, y_min = NULL, y_max = NULL)
}
\arguments{
\item{trend}{Dataframe of yearly indices produced by
\code{generate_strata_trend}}

\item{y_min}{Minimum year to plot}

\item{y_max}{Maximum year to plot}
}
\value{
List of ggplot objects, each entry being a plot
  of a stratum trend
}
\description{
Generates the trend plot for each stratum modelled.
}
