% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Wrangle data to use for modelling input (STAN TIDY)}
\usage{
prepare_data(
  strata_data = NULL,
  species = NULL,
  min_year = NULL,
  max_year = NULL,
  min_n_routes = 3,
  min_max_route_years = 3,
  min_mean_route_years = 1,
  strata_rem = NULL
)
}
\arguments{
\item{strata_data}{List. Stratified data returned by \code{stratify()}}

\item{species}{Character. English name of the species to run}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{min_n_routes}{Numeric. Required minimum routes per strata where species
has been observed. Defaults to 3}

\item{min_max_route_years}{Required minimum number of years with non-zero
observations of species on at least 1 route. Defaults to 3}

\item{min_mean_route_years}{Required minimum average of years per route with
the species observed. Defaults to 1.}

\item{strata_rem}{Character vector. Strata to remove from analysis.}
}
\value{
List of data to be used for modelling as input to \code{run_model()}.
}
\description{
Subset raw BBS data by selected species and and wrangle stratified data for
use as input for models.
}
\examples{
# Toy example with Pacific Wren sample data

# First, stratify the sample data

strat_data <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in a model. In this
#   toy example, we will set the minimum year as 2009 and
#   maximum year as 2018, effectively only setting up to
#   model 10 years of data.

model_data <- prepare_data(strat_data = strat_data,
                           species_to_run = "Pacific Wren",
                           min_year = 2009, max_year = 2018)
}
