% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-jags-data.R
\name{prepare_jags_data}
\alias{prepare_jags_data}
\title{Transform data to use for JAGS input}
\usage{
prepare_jags_data(data_strat, species_to_run, model, n_knots = 9)
}
\arguments{
\item{data_strat}{Large list of stratified data returned by \code{stratify()}}

\item{species_to_run}{Character string of the English name of the species to run}

\item{model}{Character strings or vector of character strings of what
species are wanting to be analysed.}

\item{n_knots}{Number of knots to be used in GAM function}
}
\value{
Large list of data to be used in JAGS
}
\description{
\code{prepare_jags_data} transforms stratified data for use as input
  to run JAGS models.
}
\examples{
data.jags <- prepareJAGSdata(data_strat, species_to_run = "Barn Swallow", model = "standard")
data.jags <- prepareJAGSdata(stratify(bbs.data), species_to_run = "Bufflehead", model = "gam", n_knots = 9)

}
