% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-model.R
\name{prepare_model}
\alias{prepare_model}
\title{Prepare model parameters}
\usage{
prepare_model(
  prepped_data,
  model,
  model_variant = "hier",
  model_file = NULL,
  spatial_data = NULL,
  use_pois = FALSE,
  heavy_tailed = TRUE,
  n_knots = NULL,
  basis = "mgcv",
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_CV = FALSE,
  set_seed = NULL,
  quiet = FALSE
)
}
\arguments{
\item{prepped_data}{List. Prepared data generated by \code{prepare_data()}.}

\item{model}{Character. Type of model to use, must be one of "first_diff"
(First Differences), "gam" (General Additive Model), "gamye" (General
Additive Model with Year Effect), or "slope" (Slope model).}

\item{model_variant}{Character. Model variant to use, must be one of
"nonhier" (Non-hierarchical), "hier" (Hierarchical), or "spatial"
(Spatially explicit).}

\item{model_file}{Character. Location of a custom Stan model file to use.}

\item{spatial_data}{List. Spatial neighbourhoods generated by \code{prepare_spatial()}.}

\item{use_pois}{Logical. Whether to use an Over-Dispersed Poisson model
(\code{TRUE}) or an Negative Binomial model (\code{FALSE}, default).}

\item{heavy_tailed}{Logical. Whether extra-Poisson error distributions should
be modelled as a t-distribution, with heavier tails than the standard
normal distribution. Default is \code{TRUE}. Recent results suggest this is best
even though it requires much longer convergence times. Can only be set to
FALSE with Poisson models (i.e. \code{use_pois = TRUE}).}

\item{n_knots}{Numeric. Number of knots for "gam" and "gamye" models}

\item{basis}{Character. Basis function to use for GAM smooth, one of
"original" or "mgcv". Default is "original", the same basis used in Smith
and Edwards 2020. "mgcv" is an alternate that uses the "tp" basis from the
package mgcv (also used in brms, and rstanarm). If using the "mgcv" option,
the user may want to consider adjusting the prior distributions for the
parameters and their precision.}

\item{calculate_nu}{Logical. Whether to calculate the \code{nu} parameter as
a factor of \code{gamma(2, 0.1)}. Default is \code{FALSE}.}

\item{calculate_log_lik}{Logical. Whether to calculate point-wise
log-likelihood of the data given the model. Default is \code{FALSE}.}

\item{calculate_CV}{Logical. Whether to use cross validation.}

\item{set_seed}{Numeric. If \code{NULL} (default) no seed is set. Otherwise an
integer number to be used with \code{withr::with_seed()} internally to ensure
reproducibility.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}
}
\value{
A list containing the pared model data (\code{model_data}), initialization
parameters (\code{init_values}), meta data for the analysis (\code{meta_data}), meta
data for the strata (\code{meta_strata}) and prepared data counts from
\code{prepare_data()} (\code{raw_data}).
}
\description{
Prepare model parameters
}
\details{
There are two ways you can customize
the model run. The first is to supply a custom \code{model_file} created with the
\code{copy_model_file()} function and then edited by hand.

Second, you can edit or overwrite the initialization parameters
(\code{init_values}) in the output of \code{prepare_model()} to customize the \code{init}
supplied to \code{cmdstanr::sample()}. You can supply these parameters in anyway
that \code{cmdstanr::sample()} accepts the \code{init} argument.

See the \href{https://steffilazerte.ca/bbsBayes/articles/models.html}{models article} for more
advanced examples and explanations.
}
\examples{

s <- stratify(by = "bbs_cws", sample_data = TRUE)
p <- prepare_data(s)
pm <- prepare_model(p, model = "first_diff", model_variant = "hier")
}
