% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-plot-data.R
\name{prepare_plot_data}
\alias{prepare_plot_data}
\title{Prepare data to be plotted}
\usage{
prepare_plot_data(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\description{
Subsets area weights and extracts n values from JAGS samples
}
\keyword{internal}
