% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-spatial.R
\name{prepare_spatial}
\alias{prepare_spatial}
\title{Define neighbouring strata for spatial analyses}
\usage{
prepare_spatial(
  strata_map,
  prepped_data,
  voronoi = FALSE,
  nearest_fill = FALSE,
  island_link_dist_factor = 1.2,
  buffer_type = "buffer",
  buffer_dist = 10000,
  add_map = NULL,
  quiet = FALSE
)
}
\arguments{
\item{strata_map}{sf data frame. Map of the strata in (MULTI)POLYGONs or
(MULTI)POINTs. Must have column "strata_name" matching strata output from
\code{prepare_data()}.}

\item{prepped_data}{List. Output of \code{prepare_data()}.

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}

\item{voronoi}{Logical. Whether or not to use Voroni method for polygons.
(Must use Voronoi method for points).}

\item{nearest_fill}{Logical. For strata with no neighbours, whether or not to
fill in by centroids of the 2 nearest neighbours when \strong{not} using the
Voronoi method.}

\item{island_link_dist_factor}{Numeric. Distances within a factor of this
amount are considered nearest strata neighbours. Used when linking
otherwise isolated islands of strata, when \strong{not} using the
Voronoi method.}

\item{buffer_type}{Character. Which buffer type to use when using the Voronoi
method. Must be one of \code{buffer} or \code{convex_hull}. See Details for
specifics.}

\item{buffer_dist}{Numeric. Distance to buffer and link the strata if not
connected when using the Voronoi method. Units are that of
\code{sf::st_crs(strata_map)}. This is the \emph{starting} distance if \code{buffer_type = "buffer"} or the final distance if \code{buffer_type = "convex_hull"}. See
Details.}

\item{add_map}{sf spatial object. Spatial data to add to map output.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}
}
\value{
List containing samples, nodes, adjacent matrix and map visualizing
the matrix
}
\description{
Given a spatial data frame containing polygons or points to outline strata,
a neighbourhood matrix is identified for use in \code{run_model()}
}
\details{
When using the Voronoi method, a buffer is used to fill around and link
strata together. If the \code{buffer_type} is \code{buffer}, \code{buffer_dist} is the
starting distance over which to buffer. If not all strata are linked, this
distance is increased by 10\% and applied again, repeating until all strata
are linked. If \code{buffer_type} is \code{convex_hull}, then a convex hull is used to
link up the strata before appling a buffer at a distance of \code{buffer_dist}.
Note that all distances are in the units of \code{sf::st_crs(strata_map)}.
}
\examples{
bbs_data <- stratify(by = "bbs_cws", species = "Connecticut Warbler")

model_data <- prepare_data(bbs_data, min_max_route_years = 2)

map <- load_map("bbs_cws")

sp <- prepare_spatial(map, model_data)

}
