% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{remove_bbs_data}
\alias{remove_bbs_data}
\title{Remove BBS data from cache}
\usage{
remove_bbs_data(level, release, cache_dir = FALSE)
}
\arguments{
\item{level}{Character. Data to remove, one of "all", "state", or "stop"}

\item{release}{Character/Numeric. Data to remove, one of "all", 2020, or 2022}

\item{cache_dir}{Logical. Remove entire cache directory (and all data
contained therein)}
}
\value{
Nothing.
}
\description{
Remove all or some of the data downloaded via \code{fetch_bbs_data()}
}
\examples{

\dontrun{
# Remove everything
remove_bbs_data(cache_dir = TRUE)

# Remove all files (but not the dir)
remove_bbs_data(level = "all", release = "all")

# Remove all 'stop' data
remove_bbs_data(level = "stop", release = "all")

# Remove all 2020 data
remove_bbs_data(level = "all", release = 2020)

# Remove 2020 stop data
remove_bbs_data(level = "stop", release = 2020)
}

}
