% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{remove_cache}
\alias{remove_cache}
\title{Remove bbsBayes cache}
\usage{
remove_cache(type = "bbs_data", level, release)
}
\arguments{
\item{level}{Character. Data to remove, one of "all", "state", or "stop"}

\item{release}{Character/Numeric. Data to remove, one of "all", 2020, or 2022}

\item{cache_dir}{Logical. Remove entire cache directory (and all data
contained therein)}
}
\value{
Nothing.
}
\description{
Remove all or some of the data downloaded via \code{fetch_bbs_data()} as well as
model executables created by \code{cmdstanr::cmdstan_model()} via \code{run_model()}.
}
\examples{

\dontrun{
# Remove everything
remove_cache(type = "all")

# Remove all BBS data files (but not the dir)
remove_cache(level = "all", release = "all")

# Remove all 'stop' data
remove_cache(level = "stop", release = "all")

# Remove all 2020 data
remoremove_cacheve_bbs_data(level = "all", release = 2020)

# Remove 2020 stop data
remove_cache(level = "stop", release = 2020)

# Remove all model executables
remove_cache(type = "model")
}

}
