% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route}
\alias{route}
\title{Yearly Route Information}
\format{A data frame with 107996 rows and 35 variables:
\describe{
 \item{countrynum}{Numerical representation of country}
 \item{statenum}{Numerical representation of political state.}
 \item{Route}{Breeding Bird Survey route number.}
 \item{RouteName}{Name of the BBS route.}
 \item{Active}{Numerical representation of whether the route is currently active:
   0 for inactive, 1 for active.}
 \item{Latitude}{Latitude of the route.}
 \item{Longitude}{Longitude of the route.}
 \item{Stratum}{Stratum number that contains this route.}
 \item{BCR}{Numerical representation of geopolitical stratum.}
 \item{LandTypeID}{Type of land that occurs on this route.}
 \item{RouteTypeID}{Type of BBS route.}
 \item{RouteTypeDetailID}{}
 \item{RPID}{}
 \item{Year}{Year of the given count.}
 \item{Month}{Numerical month of the given count.}
 \item{Day}{Calendar day of the given count.}
 \item{ObsN}{Unique number corresponding to the observer that conducted the route.}
 \item{TotalSpp}{Total species observed on the route for the run.}
 \item{StartTemp}{Temperature at the start of the route run.}
 \item{EndTemp}{Temperature at the end of the route run.}
 \item{TempScale}{Scale used for temperature report.}
 \item{StartWind}{Wind speed at the start of the route run in UNITS.}
 \item{EndWind}{Wind speed at the end of the route run in UNITS.}
 \item{StartSky}{Numeric representation of cloud cover at the start of the route run.}
 \item{EndSky}{Numeric representation of cloud cover at the end of the route run.}
 \item{StartTime}{Time the route run was started.}
 \item{EndTime}{The the route run was ended.}
 \item{Assistant}{Whether an assistant was used, 1 = YES, 0 = NO, NULL = NA}
 \item{RunType}{Type of the run.}
 \item{route}{Breeding Bird Survey route number.}
 \item{state}{Numerical representation of political state.}
 \item{rt.uni}{Combined state-route}
 \item{rt.uni.y}{Combined state-route-year}
 \item{runyear}{Year of the route run.}
}}
\usage{
route
}
\description{
A dataset containing information about each route run per year.
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
