% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{route_sample}
\alias{route_sample}
\title{Sample route data per year run (Pacific Wren only)}
\format{
An object of class \code{data.frame} with 4875 rows and 32 columns.
}
\usage{
route_sample
}
\description{
A dataset containing data for each route run per year.
NOTE: This only contains data for Pacific Wren, not the full data set.
The full count set is obtained via the function \code{fetch_bbs_data}.
The data is obtained from the United States Geological Survey and is
subject to change as new data is added each year. See Details for citation.
}
\details{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding, V. Aponte and M-A.R. Hudson. 2019.
North American Breeding Bird Survey Dataset 1966 - 2018, version 2018.0.
U.S. Geological Survey, Patuxent Wildlife Research Center. https://doi.org/10.5066/P9HE8XYJ.
}
\keyword{datasets}
