% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{save_model_run}
\alias{save_model_run}
\title{Save output of run_model()}
\usage{
save_model_run(model_output, path = NULL, quiet = FALSE)
}
\arguments{
\item{model_output}{List. Output of \code{run_model()}.}

\item{path}{Character. Optional file path to use for saved data. Defaults to
the file path used for the original run}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}
}
\value{
Nothing. Creates .rds file at \code{path}.
}
\description{
This function closely imitate \code{cmdstanr::save_object()} but saves the
entire model output oject from \code{run_model()} which contains more details
regarding stratification etc.
}
\examples{

# Create dummy model run (by default model .rds file saved by `run_model()`
m <- stratify(by = "bbs_cws", sample_data = TRUE) \%>\%
  prepare_data() \%>\%
  run_model(model = "first_diff", iter_sampling = 10,
            iter_warmup = 10, chains = 2)

# Deliberately save the file
save_model_run(m, path = "my_model.rds")
}
