% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sp}
\alias{sp}
\title{Species Information}
\format{A data frame with 748 rows and 10 variables:
\describe{
 \item{seq}{}
 \item{aou}{Numerical code for the given species as assigned by
   American Ornithological Union}
 \item{english}{English name of the bird species.}
 \item{french}{French name of the bird species.}
 \item{spanish}{Spanish name of the bird species.}
 \item{order}{Taxonomic order name of the bird species.}
 \item{family}{Taxonomic family name of the bird species.}
 \item{genus}{Taxonomic genus name of the bird species.}
 \item{species}{Taxonomic species name of the bird species.}
 \item{sp.bbs}{Numerical code for the given species as assigned by
   Breeding Bird Survey.}
}}
\usage{
sp
}
\description{
A dataset containing information on all speices covered
by the American Ornithological Union
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
