% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_forms}
\alias{species_forms}
\title{Species forms}
\format{
An object of class \code{data.frame} with 13 rows and 5 columns.
}
\usage{
species_forms
}
\description{
Data frame of species forms that will be combined if \code{combine_species_forms}
is \code{TRUE} in \code{stratify()}.
}
\keyword{datasets}
