% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{st.areas}
\alias{st.areas}
\title{Strata Information}
\format{A data frame with 201 rows and 13 variables:
\describe{
 \item{State}{Full name of the state or province.}
 \item{country}{Two-letter country code.}
 \item{prov}{Two letter state or province code.}
 \item{St_12}{Combined two letter codes of country-province/state-
   bcr number}
 \item{bcr}{Numerical representation of geopolitical stratum.}
 \item{SUM_Count_}{}
 \item{MIN_BCR}{}
 \item{Area}{Area of the stratum.}
 \item{St_num}{}
 \item{strat.name}{Name of the stratum.}
 \item{countrynum}{Numerical representation of the country.}
 \item{RegionCode}{Numerical representation of the region.}
 \item{state}{}
}}
\usage{
st.areas
}
\description{
A dataset containing information about all BBS strata.
}
\details{
Please use the following citation for BBS data:

Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson.
2018. North American Breeding Bird Survey Dataset 1966 - 2017,
version 2017.0. U.S. Geological Survey, Patuxent Wildlife Research
Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
