% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify raw Breeding Bird Survey data}
\usage{
stratify(bbs_data, stratify_by = "bbs")
}
\arguments{
\item{bbs_data}{Large list of raw BBS data. Can be obtained using
the \code{fetchBBSdata()} function.}

\item{stratify_by}{String argument of stratification type. Defaults
to "bbs", which is the standard BCR X State used in BBS data analysis.
Other options are "state", "bcr", "latlong"}
}
\value{
Large list (4 elements) containing stratified point count data,
route data, and species data, as well as a data frame containing area
information about each strata.
}
\description{
Assigns each bird count data point and each route a strata
  based on its geographic location and the stratification
  as specified by the user.
}
\examples{
data.strat <- stratify(bbs_data = data, stratify_by = "bbs")
data.strat <- stratify(fetchBBSdata())

}
