% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Calculate Watanabe–Akaike information criterion}
\usage{
waic(jags_data = NULL, jags_mod = NULL)
}
\arguments{
\item{jags_data}{Data prepared by \code{prepare_jags_data}, used
for input to the JAGS model}

\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Double precision numerical value
}
\description{
\code{waic} returns the Watanabe–Akaike information criterion (WAIC) for the
  supplied model. WAIC is a generalization of the Akaike information
  criterion onto singular statistical model.
}
\details{
NOTE: in order to calculated WAIC, the model MUST track the parameter
  "lambda". In species that are data-rich, such as Wood Thrush,
  this produces extremely large JAGS objects, and takes up a considerable
  amount of memory when simulating with \code{run_model}

See examples for details.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

  # Output WAIC
  waic(jags_data = prepped_data,
       jags_mod = mod)
}

}
