% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbs_strata}
\alias{bbs_strata}
\title{List of included strata}
\format{
\subsection{\code{bbs_strata}}{

A list of 5 data frames

Contains \code{bbs_usgs}, \code{bbs_cws}, \code{bcr}, \code{latlong} and \code{prov_state}
}
}
\usage{
bbs_strata
}
\description{
List of strata included in bbsBayes. Each list item contains a data frame
describing the strata for that stratification (name, area, country, etc.)
}
\keyword{datasets}
