% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bbsBayes-deprecated.R
\name{bbsBayes-deprecated}
\alias{bbsBayes-deprecated}
\alias{generate_map}
\alias{geofacet_plot}
\title{bbsBayes deprecated functions}
\usage{
generate_map(...)

geofacet_plot(...)
}
\arguments{
\item{...}{Original function arguments}
}
\description{
\code{plot_map()} supersedes \code{generate_map()}

\code{plot_geofacet()} supersedes \code{geofacet_plot()}
}
\seealso{
\link{bbsBayes-defunct}
}
