% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa_common_docs.R
\name{common_docs}
\alias{common_docs}
\title{Common arguments and documentation for various functions}
\arguments{
\item{level}{Character. Which type of BBS counts to fetch, "state" or "stop".
Defaults to "state".}

\item{release}{Numeric. Yearly release. Options are 2022 (default including
data through 2021 field season) or 2020 (including data through 2019)}

\item{sf}{Spatial data frame. PLACEHOLDER}

\item{model_output}{List. Output of \code{run_model()}.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{indices}{List. Indices generated by \code{generate_indices()}}

\item{prepped_data}{List. Output of \code{prepare_data()}.

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}
}
\description{
Common arguments and documentation for various functions
}
\section{NatureCounts account}{

All public data is available with a username/password
(\href{https://www.naturecounts.ca/nc/default/register.jsp}{sign up}
for a free NatureCounts account). However, to access private/semi-public
projects/collections you must request access. See the Access and
\code{request_id}s section for more information.
}

\keyword{internal}
