% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{convergence}
\alias{convergence}
\title{Convergence metrics}
\usage{
convergence(model_fit)
}
\arguments{
\item{model_fit}{}
}
\description{
Convergence metrics
}
\examples{

load(system.file("model_outputs", "pacific_wren_slope_BBS_fit.RData", package = "bbsBayes"))

}
