% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-index-data.R
\name{extract_index_data}
\alias{extract_index_data}
\title{Extract index data (STAN)}
\usage{
extract_index_data(model_fit, model_data, stratify_by, alt_n = "n")
}
\arguments{
\item{model_fit}{Model fit list generated by \code{run_model}}
}
\value{
List of index values produced by JAGS samples
}
\description{
Subsets area weights and extracts n values from JAGS samples
}
\keyword{internal}
