% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{fetch_bbs_data}
\alias{fetch_bbs_data}
\title{Fetch Breeding Bird Survey dataset}
\usage{
fetch_bbs_data(
  level = "state",
  release = 2022,
  force = FALSE,
  quiet = FALSE,
  compression = "none"
)
}
\arguments{
\item{level}{Character. Which type of BBS counts to fetch, "state" or "stop".
Defaults to "state".}

\item{release}{Numeric. Yearly release. Options are 2022 (default including
data through 2021 field season) or 2020 (including data through 2019)}

\item{force}{Logical. Should pre-exising BBS data be overwritten? Defaults to
FALSE}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}

\item{compression}{Character. What compression should be used to save data?
Defaults to \code{none} which takes up the most space but is the fastest to
load. Must be one of \code{none}, \code{gz}, \code{bz2}, or \code{xz} (passed to
\code{readr::write_rds()}'s \code{compress} argument).}
}
\description{
Use File Transfer Protocol (FTP) to fetch Breeding Bird Survey data from the
United States Geological Survey (USGS) FTP site. This is the raw data that is
uploaded to the site before any analyses are performed. A package-specific
directory is created on the user's computer (see documentation of
\code{rappdirs::appdir} for details of where this directory lives), and the BBS
data is saved to that directory for use by other functions. Before
downloading any data, the user must thoroughly read through the terms and
conditions of the user of the data and type the word "yes" to agree.
}
\details{
BBS \code{state} level counts provide counts beginning in 1966,
aggregated in five bins, each of which contains cumulative counts from 10
of the 50 stops along a route. In contrats BBS \code{stop} level counts provides
stop-level data beginning in 1997, which includes counts for each stop
along routes individually. Note that stop-level data is not currently
supported by the modelling utilities in bbsBayes.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}

fetch_bbs_data(force = TRUE)
fetch_bbs_data(level = "stop", force = TRUE)
fetch_bbs_data(release = 2020, force = TRUE)
fetch_bbs_data(release = 2020, level = "stop", force = TRUE)
\dontshow{\}) # examplesIf}
}
