% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{fetch_bbs_data}
\alias{fetch_bbs_data}
\title{Fetch Breeding Bird Survey dataset}
\usage{
fetch_bbs_data()
}
\value{
Large list (3 elements) containing point count data, route data, and species data.
}
\description{
\code{fetch_bbs_data} uses File Transfer Protocol (FTP) to fetch Breeding Bird
 Survey data from the United States Geological Survey (USGS) FTP site. This
 is the raw data that is uploaded to the site before any analyses are performed.
}
\examples{

\dontrun{

# Be sure to assign the output of fetch_bbs_data() to a variable.
# This function downloads a lot of data, and if the output is not
# captured in a variable, R attempts to dump it into the console
# which tends to get highly bogged down.

bbs_data <- fetch_bbs_data()

# USGS updates their data up to twice a year, so one potential
# option is to save the bbs_data to a file so that you do not
# have to run fetch_bbs_data() every new session.

bbs_data <- fetch_bbs_data()
save(bbs_data, file = "bbs_data.RData")
}

}
