% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-data.R
\name{fetchBBSdata}
\alias{fetchBBSdata}
\title{Fetch Breeding Bird Survey dataset}
\usage{
fetchBBSdata()
}
\description{
\code{fetchBBSdata} uses File Transfer Protocol (FTP) to fetch Breeding Bird
 Survey data from the United States Geological Survey (USGS) FTP site. This
 is the raw data that is uploaded to the site before any analyses are performed.
}
\details{
@return Large list (5 elements) containing point count data, route data, weather
   data, and species data. This is a very large file.
 @export
}
