% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-indices.R
\name{generate_cont_indices}
\alias{generate_cont_indices}
\title{Generate a dataframe of the continental yearly indices}
\usage{
generate_cont_indices(jags_mod = NULL, quantiles = c(0.025, 0.05, 0.25,
  0.75, 0.95, 0.975), alternate_n = "n")
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{quantiles}{vector of quantiles to be sampled from the posterior distribution Defaults to c(0.025,0.05,0.25,0.5,0.75,0.95,0.975)}

\item{alternate_n}{text string indicating the name of the alternative approach to calculating an annual index, Default is "n"}
}
\value{
List of 6 objects:
  \item{data_summary}{dataframe with the following columns}
  \item{Year}{Year of particular index}
  \item{Region}{Region name}
  \item{Index}{Strata-weighted count index}
  \item{additional columns for each of the values in quantiles}{quantiles of the posterior distribution}
  \item{samples}{array of all samples from the posterior distribution}
  \item{area-weights}{data frame of the strata names and area weights used to calculate the continental estimates}
  \item{y_min}{first year used in the model, scale 1:length of time-series}
  \item{y_max}{last year used in the model, scale 1:length of time-series}
  \item{startyear}{first year used in the model, scale 1966:2018}
}
\description{
\code{generate_cont_indices} creates a data frame of the strata-weighed
  continental indices by year. This data frame can then be used to
  generate a population trajectory of the species.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_index <- generate_cont_indices(jags_mod = mod)
}

}
