% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-indices.R
\name{generate_cont_indices}
\alias{generate_cont_indices}
\title{Generate a dataframe of the continental yearly indices}
\usage{
generate_cont_indices(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Data frame of 4 variables:
  \item{Year}{Year of particular index}
  \item{Index}{Strata-weighted count index}
  \item{Q25}{2.5\% quantile of mean strata-weighted indices}
  \item{Q975}{97.5\% quantile of mean strata-weighted indices}
}
\description{
\code{generate_cont_indices} creates a data frame of the strata-weighed
  continental indices by year. This data frame can then be used to
  generate a population trajectory of the species.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_index <- generate_cont_indices(jags_mod = mod)
}

}
