% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-trend.R
\name{generate_cont_trend}
\alias{generate_cont_trend}
\title{Generate a dataframe of the continental trend}
\usage{
generate_cont_trend(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Data frame of 4 variables:
  \item{Year}{Year of particular index}
  \item{Index}{Strata-weighted count index}
  \item{Q25}{2.5\% quantile of mean strata-weighted indices}
  \item{Q975}{97.5\% quantile of mean strata-weighted indices}
}
\description{
\code{generate_cont_trend} creates the strata-weighted continental
  trend and returns a data frame. This is useful if you would like
  to plot the year-to-year trend using your own plot recipes.
}
