% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-cont-trend.R
\name{generate_cont_trend}
\alias{generate_cont_trend}
\title{Generate the continental trend}
\usage{
generate_cont_trend(indices = NULL, min_year = NULL, max_year = NULL)
}
\arguments{
\item{indices}{Continental indices generated by \code{generate_cont_indices}}

\item{min_year}{Minimum year to calculate trends from}

\item{max_year}{Maximum year to calculate trends to}
}
\value{
Numeric percentage of trend
}
\description{
\code{generate_cont_trend} calculates the geometric mean annual changes in population size.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_index <- generate_cont_indices(jags_mod = mod)

#Output the trend
generate_cont_trend(indices = cont_index)
}

}
