% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_map.R
\name{generate_map}
\alias{generate_map}
\title{Generate a map of trends by strata.}
\usage{
generate_map(indices = NULL, stratify_by = NULL, min_year = NULL,
  max_year = NULL)
}
\arguments{
\item{indices}{Dataframe of yearly stratum indices produced by
\code{generate_strata_indices}}

\item{stratify_by}{How was the data stratified?}

\item{min_year}{Minimum year to plot}

\item{max_year}{Maximum year to plot}
}
\value{
spplot object
}
\description{
\code{generate_map} allows you to generate a colour-coded map of species trends
  for each strata. Given indices generated by \code{generate_strata_indices}, this
  function will calculate percent change for either the entire survey time or a
  specified amount of time and will then shade in each stratum based on the percent
  change in that stratum.
}
\examples{

\dontrun{
}
}
