% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-map.R
\name{generate_map}
\alias{generate_map}
\title{Generate a map of trends by strata.}
\usage{
generate_map(trend = NULL, stratify_by = NULL)
}
\arguments{
\item{trend}{Dataframe of strata trends produced by
\code{generate_strata_trends}}

\item{stratify_by}{How was the data stratified?}
}
\value{
spplot object
}
\description{
\code{generate_map} allows you to generate a colour-coded map of species trends
  for each strata. Given trends generated by \code{generate_strata_trends}, this
  function will shade in each stratum based on the percent
  change in that stratum.
}
\examples{

\dontrun{
# Set stratification type for future use
stratification <- "bcr"

# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = stratification)
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the indices for each strata
strata_index <- generate_strata_indices(jags_mod = mod)

# Get the data frame of trends by strata
trend <- generate_strata_trends(indices = strata_index)

# Obtain a map of the trends by each strata
map <- generate_map(trend = trend, stratify_by = stratification)
}
}
