% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-map.R
\name{generate_map}
\alias{generate_map}
\title{Generate a map of trends by strata.}
\usage{
generate_map(trend = NULL, select = F, stratify_by = NULL,
  slope = FALSE)
}
\arguments{
\item{trend}{Dataframe of strata trends produced by
\code{generate_strata_trends} or \code{generate_regional_trends(..., regions = "stratum")}}

\item{select}{logical flag to indicate if the stratum data need to be selected out of an trends object that includes continental, national, or other region-types. Default is FALSE}

\item{stratify_by}{How were the data stratified?}

\item{slope}{Logical, if TRUE, maps values of the alternative trend metric if slope = T was used in \code{generate_strata_trends}, the slope of a log-linear regression through the annual indices. Default FALSE.}
}
\value{
spplot object
}
\description{
\code{generate_map} allows you to generate a colour-coded map of species trends
  for each strata. Given trends generated by \code{generate_strata_trends}, this
  function will shade in each stratum based on the percent
  change in that stratum.
}
\examples{

\dontrun{
# Set stratification type for future use
stratification <- "bcr"

# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = stratification)
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the indices for each strata
strata_index <- generate_strata_indices(jags_mod = mod)

# Get the data frame of trends by strata
trend <- generate_strata_trends(indices = strata_index)

# Obtain a map of the trends by each strata
map <- generate_map(trend = trend, stratify_by = stratification)
}
}
