% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-regional-indices.R
\name{generate_regional_indices}
\alias{generate_regional_indices}
\title{Generate a dataframe of the regional yearly indices}
\usage{
generate_regional_indices(jags_mod = NULL, quantiles = c(0.025, 0.05,
  0.25, 0.75, 0.95, 0.975), regions = c("stratum", "continental"),
  alternate_n = "n", startyear = NULL, max_backcast = 5)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model} using either "bbs_cws" or "bbs_usgs" stratifications}

\item{quantiles}{vector of quantiles to be sampled from the posterior distribution Defaults to c(0.025,0.05,0.25,0.5,0.75,0.95,0.975)}

\item{regions}{vector selcting regional compilation(s) to calculate. Default is "continental","stratum", options include national", "prov_state", or "bcr", for the stratifications that include areas that align with those regions.}

\item{alternate_n}{text string indicating the name of the alternative approach to calculating an annual index, Default is "n"}

\item{startyear}{Optional first year for which to calculate the annual indices if a trajectory for only the more recent portion of the time series is desired. This is probably most relevant if max_backcast is set and so trajectories for different time-periods could include a different subset of strata (i.e., strata removed)}

\item{max_backcast}{an integer indicating the maximum number of years to backcast the stratum-level estimates before the first year in which the species was observed on any route in that stratum. Default is 5. If the observed data in a given stratum do not include at least one non-zero observation of the species between \item{startyear} and the end of the time series, the stratum is dropped from the relevant regional summary. Optionally this can be set to NULL to ignore any backcasting limit (i.e., to generate annual indices for the entire time series, regardless of when the species was first observed)}
}
\value{
List of 6 objects:
  \item{data_summary}{dataframe with the following columns}
  \item{Year}{Year of particular index}
  \item{Region}{Region name}
  \item{Index}{Strata-weighted count index}
  \item{additional columns for each of the values in quantiles}{quantiles of the posterior distribution}

  \item{samples}{array of all samples from the posterior distribution}
  \item{area-weights}{data frame of the strata names and area weights used to calculate the continental estimates}
  \item{y_min}{first year used in the summary, scale 1:length of time-series}
  \item{y_max}{last year used in the summary, scale 1:length of time-series}
  \item{startyear}{first year used in the summary, scale 1966:2018}
}
\description{
\code{generate_regional_indices} creates a data frame of the strata-weighted
  regional indices by year. This data frame can then be used to
  generate a population trajectory of the species.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Wood Thrush",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_index <- generate_cont_indices(jags_mod = mod)
}

}
