% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-strata-trends.R
\name{generate_strata_trends}
\alias{generate_strata_trends}
\title{Generate a dataframe of the stratum trends}
\usage{
generate_strata_trends(indices = NULL, min_year = NULL,
  max_year = NULL)
}
\arguments{
\item{indices}{Stratum indices generated by \code{generate_strata_indices}}

\item{min_year}{Minimum year to calculate trends from}

\item{max_year}{Maximum year to calculate trends to}
}
\value{
Data frame of 2 variables:
  \item{Stratum}{Name of the stratum}
  \item{Trend}{Percent change from the minimum year to maximum year}
}
\description{
\code{generate_strata_trends} creates a data frame of trends by year,
  factored by each stratum. These trends are calculated using the indices
  generated by \code{generate_strata_indices}.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the indices for each strata
strata_index <- generate_strata_indices(jags_mod = mod)

# Get the data frame of trends by strata
trend <- generate_strata_trends(indices = strata_index)
}

}
