% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-trend-plot.R
\name{generate_trend_plot}
\alias{generate_trend_plot}
\title{Generate trend indices for a given species.}
\usage{
generate_trend_plot(jags_mod = NULL, stratify_by = NULL,
  plot_strata = NULL)
}
\arguments{
\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{stratify_by}{How was the data stratified?}

\item{plot_strata}{Which strata would you like to include in the list
of plots? Provide a list of strata to plot or the string "all"
to plot all the strata used in the analysis.}
}
\value{
List of ggplot2 objects
}
\description{
Generates yearly indices for the specified bird species
  given simulated MCMC chains.
}
\examples{

\dontrun{
}

}
