% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate-trends.R
\name{generate_trends}
\alias{generate_trends}
\title{Generate regional trends}
\usage{
generate_trends(
  indices,
  min_year = NULL,
  max_year = NULL,
  quantiles = c(0.025, 0.05, 0.25, 0.75, 0.95, 0.975),
  slope = FALSE,
  prob_decrease = NULL,
  prob_increase = NULL,
  Min_year,
  Max_year
)
}
\arguments{
\item{indices}{List. Indices generated by \code{generate_indices()}.}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{quantiles}{Numeric vector. Quantiles to be sampled from the posterior
distribution. Defaults to \code{c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975)}}

\item{slope}{Logical. Whether to calculate an alternative trend metric, the
slope of a log-linear regression through the annual indices. Default
\code{FALSE}.}

\item{prob_decrease}{Numeric vector. Percent-decrease values for which to
optionally calculate the posterior probabilities (see Details). Default is
\code{NULL} (do not calculate).}

\item{prob_increase}{Numeric vector. Percent-increase values for which to
optionally calculate the posterior probabilities (see Details). Default is
\code{NULL} (do not calculate).}

\item{Min_year}{Deprecated. Use \code{min_year} instead}

\item{Max_year}{Deprecated. Use \code{max_year} instead}
}
\value{
A list containing the trends (\code{trends}), meta data for the analysis
(\code{meta_data}), meta data for the strata (\code{meta_strata}) and prepared data
counts from \code{prepare_data()} (\code{raw_data}).

\code{trends} is a data frame with one row for each region in the input
\code{indices}. It has the following columns:
\itemize{
\item \code{start_year} - First year of the trend
\item \code{end_year} - Last year of the trend
\item \code{region} - Region name
\item \code{region_type} - Type of region
\item \code{strata_included} - Strata \emph{potentially} included in the annual index
calculations
\item \code{strata_excluded} - Strata \emph{potentially} excluded from the annual index
calculations because they have no observations of the species in the first
part of the time series, see arguments \code{max_backcast} and \code{start_year}
\item \code{trend} - Estimated median annual percent change over the trend
time-period according to end point comparison of annual indices for the
\code{start_year} and the \code{end_year}
\item \code{trend_q_XXX} - Trend estimates by different quantiles
\item \code{percent_change} - Median overall estimate percent change over the trend
time-period
\item \code{percent_change_q_XXX} - Percent change by different quantiles
\item \code{slope_trend} - Estimated median annual percent change over the trend
time-period, according to the slope of a linear regression through the
log-transformed annual indices. (Only if \code{slope = TRUE})
\item \code{slope_trend_q_XXX} - Slope-based trend estimates by different quantiles.
(Only if \code{slope = TRUE})
\item \code{width_of_95_percent_credible_interval} - Width (in percent/year) of the
credible interval on the trend calculation. Calculated for the widest
credible interval requested in via \code{quantiles}. Default is 95 percent CI
(i.e., \code{trend_q_0.975} - \code{trend_q_0.025})
\item \code{width_of_95_percent_credible_interval_slope} - Width (in percent/year)
of the credible interval on the slope-based trend calculation. Calculated
for the widest credible interval requested in via \code{quantiles}. Default is
95 percent CI (i.e., \code{slope_trend_q_0.975} - \code{slope_trend_q_0.025}). (Only
if \code{slope = TRUE})
\item \code{prob_decrease_XX_percent} - Proportion of the posterior distribution of
\code{percent_change} that is below the percentage values in
\code{prob_decrease} (if non-\code{Null})
\item \code{prob_increase_XX_percent} - Proportion of the posterior distribution of
\code{percent_change} that is above tthe percentage values in
\code{prob_increase} (if non-\code{Null})
\item \code{rel_abundance} - Mean annual index value across all years. An estimate
of the average relative abundance of the species in the region. Can be
interpreted as the predicted average count of the species in an average
year on an average route by an average observer, for the years, routes, and
observers in the existing data
\item \code{obs_rel_abundance} - Mean observed annual count of birds across all
routes and all years. An alternative estimate of the average relative
abundance of the species in the region. For composite regions (i.e.,
anything other than stratum-level estimates) this average count is
calculated as an area-weighted average across all strata included.
\item \code{n_routes} - Number of BBS routes that contributed data for this
species and region for all years in the selected time-series, i.e., all
years since \code{start_year}
\item \code{mean_n_routes} - Mean number of BBS routes that contributed data for
this species, region, and year
\item \code{n_strata_included} - The number of strata included in the region
\item \code{backcast_flag} - Approximate annual average proportion of the covered
species range that is free of extrapolated population trajectories. e.g.,
if 1.0, data cover full time-series; if 0.75, data cover 75 percent of
time-series. Only calculated if \code{max_backcast != NULL}.
}
}
\description{
Generates trends for continent and strata and optionally for countries,
states/provinces, or BCRs from analyses run on the stratifications that
support these composite regions. Calculates the geometric mean annual changes
in population size for composite regions.
}
\details{
The posterior probabilities can be calculated for a percent-decrease
(\code{prob_decrease}) and/or percent-increase (\code{prob_increase}) if desired.
These calculate the probability that the populationhas decreased/increased
by at least the amount specified.

For example, a \code{prob_increase = 100} would result in the calculation of the
probability that the population has increased by more than 100\% (i.e.,
doubled) over the period of the trend.

Alternatively, a \code{prob_decrease = 50} would result in the calculation of
the probability that the population has decreased by more than 50\% (i.e.,
less than half of the population remains) over the period of the trend.
}
\examples{

# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices#'
i <- generate_indices(pacific_wren_model)

# Now, generate the trends
t <- generate_trends(i)


}
