% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_convergence}
\alias{get_convergence}
\title{Convergence metrics}
\usage{
get_convergence(model_output, variables = NULL)
}
\arguments{
\item{model_output}{List. Model output generated by \code{run_model()}.}

\item{variables}{Character vector. Specific variables (e.g.,
\code{"strata_raw[1]"}) or variable types (e.g., \code{"strata_raw"}) for which to
calculate metrics. If \code{NULL} (default) all variables are returned.}
}
\value{
Data frame of convergence metrics for all model variables. Contains
\verb{variable_type, }variable\verb{, }ess_bulk\verb{, }ess_tail\verb{, and }rhat`
}
\description{
Convergence metrics
}
\examples{

# Using the example model for Pacific Wrens

conv <- get_convergence(pacific_wren_model)
conv <- get_convergence(pacific_wren_model, variables = "strata_raw")
conv <- get_convergence(pacific_wren_model, variables = "strata_raw[9]")
}
