% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-final-values.R
\name{get_final_values}
\alias{get_final_values}
\title{Get final values of JAGS model}
\usage{
get_final_values(model = NULL)
}
\arguments{
\item{model}{Model object returned by \code{run_model}}
}
\value{
List of final values of monitored parameters.
}
\description{
\code{get_final_values} returns the final values of all parameters of
  the model created by \code{run_model} as a list. This function would mostly be used
  in conjunction with \code{run_model} to provide initial values.
}
