% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-mcmc-list.R
\name{get_mcmc_list}
\alias{get_mcmc_list}
\title{Obtain MCMC list from jagsUI object}
\usage{
get_mcmc_list(jags_mod = NULL)
}
\arguments{
\item{jags_mod}{JAGS object returned by \code{run_model}}
}
\value{
List containing:
\item{mcmc_list}{Traditional "coda" samples produced by rjags}
\item{sims_list}{Cleaned up posterior samples produced by jagsUI}
}
\description{
\code{get_mcmc_list} will return both the mcmc.list object
  and the sims.list object from jagsUI. mcmc.list is the usual
  MCMC samples generated by the rjags library, and sims.list
  is just a "cleaned up" version of mcmc.list produced by
  the jagsUI library.
}
\examples{

\dontrun{
  # First generate some samples from a JAGS model
  jags_mod <- run_model(jags_data = jags_data)

  # Get the posterior samples from that model
  samples <- get_mcmc_list(jags_mod = jags_mod)
}

}
