% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_model_vars}
\alias{get_model_vars}
\title{Get model variables}
\usage{
get_model_vars(model_output, all = FALSE)
}
\arguments{
\item{model_output}{List. Model output generated by \code{run_model()}.}

\item{all}{Logical. Whether or not to return \strong{all}, specific variables
(e.g., \code{strata_raw[1]} or just variable types (e.g., \code{strata_raw}).
Defaults to \code{FALSE} (variable types only).}
}
\value{
A character vector of all model variable types.
}
\description{
Returns the basic model variables types (note that most variables have
different iterations for each strata and each year).
}
\examples{

# Using the example model for Pacific Wrens...

# List variable types
get_model_vars(pacific_wren_model)

# List all variables
get_model_vars(pacific_wren_model, all = TRUE)

}
