% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence.R
\name{get_summary}
\alias{get_summary}
\title{Return the \code{cmdstanr} summary}
\usage{
get_summary(model_output, variables = NULL)
}
\arguments{
\item{model_output}{List. Output of \code{run_model()}.}

\item{variables}{Character vector. Specific variables (e.g.,
\code{"strata_raw[1]"}) or variable types (e.g., \code{"strata_raw"}) for which to
calculate metrics. If \code{NULL} (default) all variables are returned.

Use \verb{@inheritParams common_docs} to include the above in any function
documentation with a matching argument (will only include matching args)}
}
\value{
A data frame of model summary statistics
}
\description{
Extract and return the model summary using \code{cmdstanr::summary()}.
}
\examples{

# Using the example model for Pacific Wrens

get_summary(pacific_wren_model)
get_summary(pacific_wren_model, variables = "strata_raw")
get_summary(pacific_wren_model, variables = "strata_raw[9]")

}
