% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-bbs-data.R
\name{have_bbs_data}
\alias{have_bbs_data}
\title{Check whether BBS data exists}
\usage{
have_bbs_data(level = "state", release = 2022, quiet = FALSE)
}
\arguments{
\item{level}{Character. Which type of BBS counts to fetch, "state" or "stop".
Defaults to "state".}

\item{release}{Numeric. Yearly release. Options are 2022 (default including
data through 2021 field season) or 2020 (including data through 2019)}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}
}
\value{
\code{TRUE} if the data is found, \code{FALSE} otherwise
}
\description{
Check whether BBS data exists
}
