% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-bbs-data.R
\name{load_bbs_data}
\alias{load_bbs_data}
\title{Load Breeding Bird Survey dataset into R Session}
\usage{
load_bbs_data(level = "state", release = 2022, sample = FALSE, quiet = TRUE)
}
\arguments{
\item{level}{Character. Which type of BBS counts to fetch, "state" or "stop".
Defaults to "state".}

\item{release}{Numeric. Yearly release. Options are 2022 (default including
data through 2021 field season) or 2020 (including data through 2019)}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}
}
\value{
Large list (3 elements) consisting of:
\item{birds}{Data frame of sample bird point count data per route, per year}
\item{routes}{Data frame of sample yearly route data}
\item{species}{Sample list of North American bird species}
}
\description{
Loads the raw, unstratified BBS data into the current R session.
The data must have been previously fetched using \code{fetch_bbs_data()}.
Note that this is not necessary to run a Bayesian analysis of BBS data;
calling \code{stratify()} will do the loading for you.
}
