% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_map}
\alias{load_map}
\title{Load a map of of the geographic strata.}
\usage{
load_map(stratify_by = NULL, type = "strata")
}
\arguments{
\item{stratify_by}{Character. Stratification type. One of
"prov_state", "bcr", "latlong", "bbs_cws", "bbs_usgs".}

\item{type}{Character. "strata" or political map ("North America", "Canada"
or "US"/"USA"/"United States of America").}
}
\value{
sf polygon object
}
\description{
\code{load_map} allows you to load a simple features object (\code{sf} package)
that represents a map of the strata that can be easily plotted.
}
\examples{
# Toy example with Pacific Wren sample data
# First, stratify the sample data
strat_map <- load_map(stratify_by = "bbs_cws")

# simple plot of the map
plot(strat_map)


}
