% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-sample-data.R
\name{load_sample_data}
\alias{load_sample_data}
\title{Load Sample Breeding Bird Survey dataset into R Session}
\usage{
load_sample_data()
}
\value{
Large list (3 elements) consisting of:
\item{bird}{Data frame of sample bird point count data per route, per year}
\item{route}{Data frame of sample yearly route data}
\item{species}{Sample list of North American bird species}
}
\description{
\code{load_sample_data} returns the sample data provided by bbsBayes. The data
is returned as a list of data frames, similar to what is returned by \code{load_bbs_data}
}
\examples{

sample_data <- load_sample_data()


}
