% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lppd.R
\name{lppd}
\alias{lppd}
\title{Calculate log posterior predictive density}
\usage{
lppd(jags_data = NULL, jags_mod = NULL, pointwise = FALSE)
}
\arguments{
\item{jags_data}{Data prepared by \code{prepare_jags_data}, used
for input to the JAGS model}

\item{jags_mod}{JAGS list generated by \code{run_model}}

\item{pointwise}{If set to \code{TRUE}, a data frame is returned
that contains the pointwise LPPD for each count. Defaults
to \code{FALSE}}
}
\value{
Data frame of pointwise LPPD by count if \code{pointwise}
  is set to \code{TRUE}. Double precision numerical value of LPPD if
  \code{pointwise} is set to \code{FALSE}.
}
\description{
\code{lppd} Calculate log posterior preditive density (LPPD) for the supplied
  model.
}
\details{
NOTE: in order to calculated LPPD, the model MUST track the parameter
  "lambda". In species that are data-rich, such as Barn Swallow,
  this produces extremely large JAGS objects, and takes up a considerable
  amount of memory when simulating with \code{run_model}

See examples for details.
}
\examples{
\dontrun{
  # Run a JAGS model, and be sure to track the parameter "lambda"
  jags_mod <- run_model(jags_data = prepared_data,
                        variable_names = "lambda")

  # Output LPPD
  lppd(jags_data = prepared_data,
       jags_mod = jags_mod)

  # Get dataframe of pointwise LPPD values for each count
  lppd_pointwise <- lppd(jags_data = prepared_data,
                         jags_mod = jags_mod,
                         pointwise = TRUE)
}

}
