% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-to-file.R
\name{model_to_file}
\alias{model_to_file}
\title{Save model to text file}
\usage{
model_to_file(filename = NULL, model = NULL)
}
\arguments{
\item{filename}{File name to create on disk.}

\item{model}{Model to be saved}
}
\description{
\code{model_to_file} allows you to save any of the preloaded hierarchical
  Bayesian models to a text file.
}
\examples{

\dontrun{
# Save the Standard model to a file called "standard.txt"
model_to_file(filename = "standard.txt", model = "standard")
}

}
