% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pacific_wren_model}
\alias{pacific_wren_model}
\title{Example model output}
\format{
\subsection{\code{pacific_wren_model}}{

A list output from \code{run_model()} with 4 items
\itemize{
\item \code{model_fit} - The output from \code{cmdstanr::sample()}
\item \code{meta_data} - Meta data related to the model run
\item \code{meta_strata} - Meta data related to the strata
\item \code{raw_data} - Data compiled by \code{prepare_data()}
}
}
}
\usage{
pacific_wren_model
}
\description{
Example model output from running a hierarchical first difference model
on the included sample data for Pacific Wrens.
}
\keyword{datasets}
