% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cont-indices.R
\name{plot_cont_indices}
\alias{plot_cont_indices}
\title{Generate continental plot of index trajectories.}
\usage{
plot_cont_indices(indices = NULL, y_min = NULL, y_max = NULL)
}
\arguments{
\item{indices}{Dataframe of yearly indices produced by
\code{generate_cont_indices}}

\item{y_min}{Minimum year to plot}

\item{y_max}{Maximum year to plot}
}
\value{
ggplot of continental indices
}
\description{
Generates the trajectory plot of continental indices.
}
\examples{

\dontrun{
# After generating continental indicess, plot them
c_plot <- plot_cont_indices(indices = cont_index)
print(c_plot)

# You can specify to only plot a subset of years using y_min and y_max
# Plots indices from 1990 onward
c_plot <- plot_cont_indices(indices = cont_indices, y_min = 1990)
#Plot up indicess up to the year 2000
c_plot <- plot_cont_indices(indices = cont_indices, y_max = 2000)
#Plot indicess between 1970 and 2010
c_plot <- plot_cont_indices(indices = cont_indices, y_min = 1970, y_max = 2010)

}

}
