% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cont-indices.R
\name{plot_cont_indices}
\alias{plot_cont_indices}
\title{Generate continental plot of index trajectories.}
\usage{
plot_cont_indices(indices = NULL, min_year = NULL, max_year = NULL,
  species = "", title_size = 20, axis_title_size = 18,
  axis_text_size = 16)
}
\arguments{
\item{indices}{Dataframe of yearly indices produced by
\code{generate_cont_indices}}

\item{min_year}{Minimum year to plot}

\item{max_year}{Maximum year to plot}

\item{species}{Species name to be added onto the plot}

\item{title_size}{Specify font size of plot title. Defaults to 20}

\item{axis_title_size}{Specify font size of axis titles. Defaults to 18}

\item{axis_text_size}{Specify font size of axis text. Defaults to 16}
}
\value{
ggplot of continental indices
}
\description{
Generates the trajectory plot of continental indices.
}
\examples{

\dontrun{

# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the continental indices weighted by each strata
cont_indices <- generate_cont_indices(jags_mod = mod)


# After generating continental indices, plot them
plot_cont_indices(indices = cont_indices, species = "Barn Swallow")


# You can specify to only plot a subset of years using min_year and max_year
# Plots indices from 1990 onward
c_plot <- plot_cont_indices(indices = cont_indices,
                            min_year = 1990,
                            species = "Barn Swallow")
#Plot up indicess up to the year 2000
c_plot <- plot_cont_indices(indices = cont_indices,
                            max_year = 2000,
                            species = "Barn Swallow")
#Plot indicess between 1970 and 2010
c_plot <- plot_cont_indices(indices = cont_indices,
                            min_year = 1970,
                            max_year = 2010,
                            species = "Barn Swallow")

}

}
