% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cont-trend.R
\name{plot_cont_trend}
\alias{plot_cont_trend}
\title{Generate continental trend plot.}
\usage{
plot_cont_trend(trend = NULL, y_min = NULL, y_max = NULL)
}
\arguments{
\item{trend}{Dataframe of yearly indices produced by
\code{generate_cont_trend}}

\item{y_min}{Minimum year to plot}

\item{y_max}{Maximum year to plot}
}
\value{
ggplot of continental trend
}
\description{
Generates the trend plot for continental indices.
}
\examples{

\dontrun{
# After generating continental trends, plot them
c_plot <- plot_cont_trend(trend = cont_trend)
print(c_plot)

# You can specify to only plot a subset of years using y_min and y_max
# Plots trend from 1990 onward
c_plot <- plot_cont_trend(trend = cont_trend, y_min = 1990)
#Plot up trends up to the year 2000
c_plot <- plot_cont_trend(trend = cont_trend, y_max = 2000)
#Plot trends between 1970 and 2010
c_plot <- plot_cont_trend(trend = cont_trend, y_min = 1970, y_max = 2010)

}

}
