% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-cont-trend.R
\name{plot_cont_trend}
\alias{plot_cont_trend}
\title{Generate continental trend plot.}
\usage{
plot_cont_trend(trend = NULL)
}
\arguments{
\item{trend}{Dataframe of yearly indices produced by
\code{generate_cont_trend}}
}
\value{
ggplot of continental trend
}
\description{
Generates the trend plot for continental indices.
}
