% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-indices.R
\name{plot_indices}
\alias{plot_indices}
\title{Generate plots of index trajectories by stratum}
\usage{
plot_indices(
  indices = NULL,
  ci_width = 0.95,
  min_year = NULL,
  max_year = NULL,
  title = TRUE,
  title_size = 20,
  axis_title_size = 18,
  axis_text_size = 16,
  line_width = 1,
  add_observed_means = FALSE,
  add_number_routes = FALSE
)
}
\arguments{
\item{indices}{List. List of indices of annual abundance and other results
produced by \code{generate_indices()}}

\item{ci_width}{Numeric. Quantile defining the width of the plotted credible
interval. Defaults to 0.95 (lower = 0.025 and upper = 0.975)}

\item{min_year}{Numeric. Minimum year to plot}

\item{max_year}{Numeric. Maximum year to plot}

\item{axis_title_size}{Numeric. Font size of axis titles. Defaults to 18}

\item{axis_text_size}{Numeric. Font size of axis text. Defaults to 16}

\item{line_width}{Numeric. Size of the trajectory line. Defaults to 1}

\item{add_observed_means}{Logical. Whether to include points indicating the
observed mean counts. Default \code{FALSE}. Note: scale of observed means and
annual indices may not match due to imbalanced sampling among routes}

\item{add_number_routes}{Logical. Whether to superimpose plot over a dotplot
showing the number of BBS routes included in each year. This is useful as a
visual check on the relative data-density through time because in most
cases the number of observations increases over time}

\item{title_size.}{Numeric. Font size of plot title. Defaults to 20}
}
\value{
List of ggplot objects, each entry being a plot of a stratum indices
}
\description{
Generates the indices plot for each stratum modelled.
}
\examples{

# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for modelling
d <- prepare_data(s, species = "Pacific Wren",
                  min_year = 2009,
                  max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)

# Generate national, continental, and stratum indices
i <- generate_indices(model_output = m,
                      regions = c("national", "continental", "stratum"))

# Now, plot_indices() will generate a list of plots for all regions
plots <- plot_indices(i)

# To view any plot, use [[i]]
plots[[1]]

names(plots)

# Suppose we wanted to access the continental plot. We could do so with
cont_plot <- plots[["Continental"]]

# You can specify to only plot a subset of years using min_year and max_year

# Plots indices from 2015 onward
p_2015_min <- plot_indices(i, min_year = 2015)

#Plot up indices up to the year 2017
p_2017_max <- plot_indices(i, max_year = 2017)

#Plot indices between 2011 and 2016
p_2011_2016 <- plot_indices(i, min_year = 2011, max_year = 2016)

}
