% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-map.R
\name{plot_map}
\alias{plot_map}
\title{Generate a map of trends by strata}
\usage{
plot_map(
  trends,
  slope = FALSE,
  title = TRUE,
  col_viridis = FALSE,
  species,
  stratify_by,
  select
)
}
\arguments{
\item{trends}{List. Trends generated by \code{generate_trends()}.}

\item{slope}{Logical. Whether or not to map values of the alternative trend
metric (slope of a log-linear regression) if \code{slope = TRUE} was used in
\code{generate_trends()},  through the annual indices. Default \code{FALSE}.}

\item{title}{Logical. Whether or not to include a title with species. Default
\code{TRUE}.}

\item{col_viridis}{Logical flag to use "viridis" colour-blind friendly
palette. Default is FALSE}

\item{species}{Defunct. Use \code{title} instead}

\item{stratify_by}{Defunct.}

\item{select}{Defunct.}
}
\value{
spplot object
}
\description{
\code{plot_map()} allows you to generate a colour-coded map of there percent
change in species trends for each strata.
}
\examples{
# Using the example model for Pacific Wrens...

# Generate the continental and stratum indices
i <- generate_indices(pacific_wren_model)

# Now generate trends
t <- generate_trends(i)

# Generate the map
map <- plot_map(t)

}
