% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-strata-indices.R
\name{plot_strata_indices}
\alias{plot_strata_indices}
\title{Generate plots of index trajectories by stratum}
\usage{
plot_strata_indices(indices = NULL, min_year = NULL, max_year = NULL)
}
\arguments{
\item{indices}{Dataframe of yearly indices produced by
\code{generate_strata_indices}}

\item{min_year}{Minimum year to plot}

\item{max_year}{Maximum year to plot}
}
\value{
List of ggplot objects, each entry being a plot
  of a stratum indices
}
\description{
Generates the indices plot for each stratum modelled.
}
\examples{

\dontrun{
# Run a JAGS model analysis on a species
stratified_data <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")
prepped_data <- prepare_jags_data(strat_data = stratified_data,
                                  species_to_run = "Barn Swallow",
                                  model = "slope")
mod <- run_model(jags_data = prepped_data)

#Generate the indices for each strata
strata_indices <- generate_strata_indices(jags_mod = mod)


# After generating strata indices, plot them
s_plot <- plot_strata_indices(indices = strata_indices)

# s_plot is just a list of ggplot objects, so you can access by index
print(s_plot[[1]])

# Or access by strata name, noting the underscores in place of special characters
print(s_plot[["US_FL_31"]])

# You can specify to only plot a subset of years using min_year and max_year
# Plots indices from 1990 onward
s_plot <- plot_strata_indices(indices = strata_indices, min_year = 1990)
#Plot up indicess up to the year 2000
s_plot <- plot_strata_indices(indices = strata_indices, max_year = 2000)
#Plot indicess between 1970 and 2010
s_plot <- plot_strata_indices(indices = strata_indices, min_year = 1970, max_year = 2010)
}
}
