% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Wrangle data to use for modelling input (STAN TIDY)}
\usage{
prepare_data(
  strat_data = NULL,
  species_to_run = NULL,
  model = NULL,
  heavy_tailed = TRUE,
  n_knots = NULL,
  min_year = NULL,
  max_year = NULL,
  min_n_routes = 3,
  min_max_route_years = 3,
  min_mean_route_years = 1,
  strata_rem = NULL,
  quiet = FALSE,
  sampler = "Stan",
  basis = "mgcv",
  use_pois = FALSE,
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_CV = FALSE
)
}
\arguments{
\item{strat_data}{List. Stratified data returned by \code{stratify()}}

\item{species_to_run}{Character. English name of the species to run}

\item{model}{Character. Model to use. One of \verb{"slope", "firstdiff", "gam", "gamye"}}

\item{heavy_tailed}{Logical. Whether the extra-Poisson error distribution
should be modeled as a t-distribution, with heavier tails than the standard
normal distribution. Default is \code{FALSE}, but recent results suggest users
should strongly consider setting this to \code{TRUE}, even though it requires
much longer convergence times}

\item{n_knots}{Numeric. Number of knots to be used in GAM function}

\item{min_year}{Numeric. Minimum year to keep in analysis}

\item{max_year}{Numeric. Maximum year to keep in analysis}

\item{min_n_routes}{Numeric. Required minimum routes per strata where species
has been observed. Defaults to 3}

\item{min_max_route_years}{Required minimum number of years with non-zero
observations of species on at least 1 route. Defaults to 3}

\item{min_mean_route_years}{Required minimum average of years per route with
the species observed. Defaults to 1.}

\item{strata_rem}{Character vector. Strata to remove from analysis.}

\item{quiet}{Should progress bars be suppressed?}

\item{basis}{Character. Which version of the basis-function to use for the
GAM smooth. Default is \code{"original"}, the same basis used in Smith and
Edwards 2020. \code{"mgcv"} is an alternate that uses the "tp" basis from the
\code{mgcv} package (also used in \code{brms}, and \code{rstanarm}). If using \code{"mgcv"},
the user may want to consider adjusting the prior distributions for the
parameters and their precision.}
}
\value{
List of data to be used for modelling as input to \code{run_model()}.
}
\description{
Subset raw BBS data by selected species and and wrangle stratified data for
use as input for models.
}
\examples{
# Toy example with Pacific Wren sample data

# First, stratify the sample data

strat_data <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in a model. In this
#   toy example, we will set the minimum year as 2009 and
#   maximum year as 2018, effectively only setting up to
#   model 10 years of data. We will use the "first difference
#   model.
model_data <- prepare_data(strat_data = strat_data,
                           species_to_run = "Pacific Wren",
                           model = "firstdiff",
                           min_year = 2009,
                           max_year = 2018)

# You can also specify the GAM model, with an optional number of
# knots to use for the GAM basis.
# By default, the number of knots will be equal to the floor
# of the total unique years for the species / 4
model_data <- prepare_data(strat_data = strat_data,
                           species_to_run = "Pacific Wren",
                           model = "gam",
                           n_knots = 9)


}
