% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-data.R
\name{prepare_data}
\alias{prepare_data}
\title{Filter for data quality}
\usage{
prepare_data(
  strata_data,
  min_year = NULL,
  max_year = NULL,
  min_n_routes = 3,
  min_max_route_years = 3,
  min_mean_route_years = 1,
  quiet = FALSE,
  species,
  model,
  heavy_tailed,
  n_knots,
  basis,
  sampler,
  strat_data,
  strata_rem
)
}
\arguments{
\item{strata_data}{List. Stratified data generated by \code{stratify()}}

\item{min_year}{Numeric. Minimum year to use. Default (\code{NULL}) uses first
year in data.}

\item{max_year}{Numeric. Maximum year to use. Default (\code{NULL}) uses first
year in data.}

\item{min_n_routes}{Numeric. Required minimum routes per strata where species
has been observed. Default 3.}

\item{min_max_route_years}{Numeric. Required minimum number of years with
non-zero observations of species on at least 1 route. Default 3. Only
retain strata with at least one route where the species was observed at
least once in this many years.}

\item{min_mean_route_years}{Numeric. Required minimum average of years per
route with the species observed. Default 1. Only retain strata where the
average number of years the species was observed per route is greater than
this value.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}

\item{species}{Defunct. Use \code{species} in \code{stratify()} instead}

\item{model}{Defunct. Use \code{model} in \code{prepare_model()} instead}

\item{heavy_tailed}{Defunct. Use \code{heavy_tailed} in \code{prepare_model()} instead}

\item{n_knots}{Defunct. Use \code{n_knots} in \code{prepare_model()} instead}

\item{basis}{Defunct. Use \code{basis} in \code{prepare_model()} instead}

\item{sampler}{Defunct.}

\item{strat_data}{Defunct. Use \code{strata_data} instead}

\item{strata_rem}{Defunct. Use \code{custom_strata} in \code{stratify()} instead}
}
\value{
List of prepared (meta) data to be used for modelling and further
steps.
\itemize{
\item \code{model_data} list of data formatted for use in Stan modelling
\item \code{meta_data} meta data defining the analysis
\item \code{meta_strata} data frame listing strata meta data
\item \code{raw_data} contains a data frame of summarized data used to create
\code{model_data} (just formatted more nicely)
}
}
\description{
Check and filter the stratified data by minimum required samples for
modelling, and prepare data format for use by models.
}
\examples{
# Toy example with Pacific Wren sample data

# First, stratify the sample data

s <- stratify(by = "bbs_cws", sample_data = TRUE)

# Prepare the stratified data for use in a model. In this
#   toy example, we will set the minimum year as 2009 and
#   maximum year as 2018, effectively only setting up to
#   model 10 years of data.

p <- prepare_data(s, min_year = 2009, max_year = 2018)
}
