% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-jags-data.R
\name{prepare_jags_data}
\alias{prepare_jags_data}
\title{Transform data to use for JAGS input}
\usage{
prepare_jags_data(data, species_to_run, model, n_knots = 9, ...)
}
\arguments{
\item{data}{Large list of stratified data returned by \code{stratify()}}

\item{species_to_run}{Character string of the English name of the species to run}

\item{model}{Character strings or vector of character strings of what
species are wanting to be analysed.}

\item{n_knots}{Number of knots to be used in GAM function}

\item{...}{Additional arguments}
}
\value{
Large list of data to be used in JAGS
}
\description{
\code{prepare_jags_data} transforms stratified data for use as input
  to run JAGS models.
}
\examples{

\dontrun{
# Download BBS data and stratify it
bbs_data <- fetch_bbs_data()
stratified_data <- stratify(bbs_data)

# Prepare the stratified data for use in a JAGS model.
# This particular instance prepares for the Standard BBS model.
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Spruce Grouse",
                               model = "standard")

# Prepare data for use the First Difference BBS model.
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Mallard",
                               model = "firstdifference")

# You can also specify the GAM model, with an optional number of
# knots to use for the GAM basis (defaults to 9 knots)
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Barn Swallow",
                               model = "gam",
                               n_knots = 9)

# This function accepts French bird names
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Oie des neiges",
                               model = "standard")

# Capitalization and punctuation matter (for now)
# This code will produce an error.
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Eastern whippoorwill"
                               model = "standard")
# But this code will be fine
data_jags <- prepare_jags_data(data = stratified_data,
                               species_to_run = "Eastern Whip-poor-will"
                               model = "standard")
}

}
