% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-model.R
\name{prepare_model}
\alias{prepare_model}
\title{Prepare model parameters}
\usage{
prepare_model(
  prepared_data,
  model,
  model_variant = "hier",
  model_file = NULL,
  use_pois = FALSE,
  heavy_tailed = TRUE,
  n_knots = NULL,
  basis = "mgcv",
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_cv = FALSE,
  cv_k = 10,
  cv_fold_groups = "obs_n",
  cv_omit_singles = TRUE,
  set_seed = NULL,
  quiet = FALSE
)
}
\arguments{
\item{prepared_data}{List. Prepared data generated by \code{prepare_data()} (if
\code{model-variant} is not \code{spatial}) or \code{prepare_spatial()} (if
\code{model_variant} is \code{"spatial"}).}

\item{model}{Character. Type of model to use, must be one of "first_diff"
(First Differences), "gam" (General Additive Model), "gamye" (General
Additive Model with Year Effect), or "slope" (Slope model).}

\item{model_variant}{Character. Model variant to use, must be one of
"nonhier" (Non-hierarchical), "hier" (Hierarchical; default), or "spatial"
(Spatially explicit).}

\item{model_file}{Character. Optional location of a custom Stan model file to
use.}

\item{use_pois}{Logical. Whether to use an Over-Dispersed Poisson model
(\code{TRUE}) or an Negative Binomial model (\code{FALSE}; default).}

\item{heavy_tailed}{Logical. Whether extra-Poisson error distributions should
be modelled as a t-distribution, with heavier tails than the standard
normal distribution. Default \code{TRUE}. Recent results suggest this is best
even though it requires much longer convergence times. Can only be set to
\code{FALSE} with Poisson models (i.e. \code{use_pois = TRUE}).}

\item{n_knots}{Numeric. Number of knots for "gam" and "gamye" models}

\item{basis}{Character. Basis function to use for GAM smooth, one of
"original" or "mgcv". Default is "original", the same basis used in Smith
and Edwards 2020. "mgcv" is an alternate that uses the "tp" basis from the
package mgcv (also used in brms, and rstanarm). If using the "mgcv" option,
the user may want to consider adjusting the prior distributions for the
parameters and their precision.}

\item{calculate_nu}{Logical. Whether to calculate the \code{nu} parameter as
a factor of \code{gamma(2, 0.1)}. Default \code{FALSE}.}

\item{calculate_log_lik}{Logical. Whether to calculate point-wise
log-likelihood of the data given the model. Default \code{FALSE}.}

\item{calculate_cv}{Logical. Whether to use bbsBayes' cross validation.
Note this is \strong{experimental}. See Details.}

\item{cv_k}{Numeric. The number of K folds to include (only relevant if
\code{calculate_cv = TRUE}). Default 10. Note this is \strong{experimental}.}

\item{cv_fold_groups}{Character. The data column to use when determining the
grouping level of the observations to be assigned to different fold groups.
Must be one of \code{obs_n} (default) or \code{routes} (only relevant if
\code{calculate_cv = TRUE}). Note this is \strong{experimental}. See the \href{https://steffilazerte.ca/bbsBayes/articles/models.html}{models article} for more
details.}

\item{cv_omit_singles}{Logical. Whether to omit test groups with no
replication (only relevant if \code{calculate_cv = TRUE}). Default \code{TRUE.} See
the \href{https://steffilazerte.ca/bbsBayes/articles/models.html}{models article} for more
details.}

\item{set_seed}{Numeric. If \code{NULL} (default) no seed is set. Otherwise an
integer number to be used with \code{withr::with_seed()} internally to ensure
reproducibility.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}
}
\value{
A list of prepared data.
\itemize{
\item \code{model_data} list of data formatted for use in Stan modelling
\item \code{init_values} list of initialization parameters
\item \code{folds} if \code{calculate_cv = TRUE}, a vector of k-fold groups each
observation is assigned to (otherwise \code{NULL})
\item \code{meta_data} meta data defining the analysis
\item \code{meta_strata} data frame listing strata meta data
\item \code{raw_data} contains a data frame of summarized data used to create
\code{model_data} (just formatted more nicely)
}
}
\description{
Calculate and format the prepared data for use in modelling. Different
parameters are definied for different types of models (see \code{?bbs_models} for
a list of models included in bbsBayes).
}
\details{
There are two ways you can customize the model run. The first is to supply a
custom \code{model_file} created with the \code{copy_model_file()} function and then
edited by hand.

Second, you can edit or overwrite the initialization parameters
(\code{init_values}) in the output of \code{prepare_model()} to customize the \code{init}
supplied to \code{cmdstanr::sample()}. You can supply these parameters in anyway
that \code{cmdstanr::sample()} accepts the \code{init} argument.

To implement bbsBayes' version of cross validation, set \code{calculate_cv = TRUE}. You can set up your own system for cross validation by modifying the
\code{folds} list-item in the output of \code{prepare_model()}.

See the \href{https://steffilazerte.ca/bbsBayes/articles/models.html}{models article} for more
advanced examples and explanations.
}
\examples{

s <- stratify(by = "bbs_cws", sample_data = TRUE)
p <- prepare_data(s)
pm <- prepare_model(p, model = "first_diff", model_variant = "hier")
}
