% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare-spatial.R
\name{prepare_spatial}
\alias{prepare_spatial}
\title{Define neighbouring strata for spatial analyses}
\usage{
prepare_spatial(
  prepared_data,
  strata_map,
  voronoi = FALSE,
  nearest_fill = FALSE,
  island_link_dist_factor = 1.2,
  buffer_type = "buffer",
  buffer_dist = 10000,
  add_map = NULL,
  label_size = 3,
  quiet = FALSE
)
}
\arguments{
\item{prepared_data}{List. Prepared data generated by \code{prepare_data()}.}

\item{strata_map}{\code{sf} Data Frame. \code{sf} map of the strata in
(MULTI)POLYGONs. Must have column "strata_name" matching strata output from
\code{prepare_data()}.}

\item{voronoi}{Logical. Whether or not to use Voroni method. Default \code{FALSE}.}

\item{nearest_fill}{Logical. For strata with no neighbours, whether or not to
fill in by centroids of the 2 nearest neighbours when \strong{not} using the
Voronoi method. Default \code{FALSE}.}

\item{island_link_dist_factor}{Numeric. Distances within a factor of this
amount are considered nearest strata neighbours. Used when linking
otherwise isolated islands of strata, when \strong{not} using the
Voronoi method. Default 1.2.}

\item{buffer_type}{Character. Which buffer type to use when using the Voronoi
method. Must be one of \code{buffer} (default) or \code{convex_hull}. See Details for
specifics.}

\item{buffer_dist}{Numeric. Distance to buffer and link the strata if not
connected when using the Voronoi method. Units are that of
\code{sf::st_crs(strata_map)}. This is the \emph{starting} distance if \code{buffer_type = "buffer"} or the final distance if \code{buffer_type = "convex_hull"}. Default
10000. See Details.}

\item{add_map}{\code{sf} object. Spatial data to add to map output.}

\item{label_size}{Numeric. Size of the labels on the map. For data with many
different strata it can be useful to reduce the size of the labels. Default
3.}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}
}
\value{
List of prepared (meta) data to be used for modelling and further
steps.
\itemize{
\item \code{spatial_data} list of samples, nodes, adjacent matrix and map
visualizing the matrix
\item \code{model_data} list of data formatted for use in Stan modelling
\item \code{meta_data} meta data defining the analysis
\item \code{meta_strata} data frame listing strata meta data
\item \code{raw_data} contains a data frame of summarized data used to create
\code{model_data} (just formatted more nicely)
}
}
\description{
Given the prepared data and a spatial data frame of polygons outlining
strata, identify a neighbourhood matrix for use in modelling.
}
\details{
When using the Voronoi method, a buffer is used to fill around and link
strata together. If the \code{buffer_type} is \code{buffer}, \code{buffer_dist} is the
starting distance over which to buffer. If not all strata are linked, this
distance is increased by 10\% and applied again, repeating until all strata
are linked. If \code{buffer_type} is \code{convex_hull}, then a convex hull is used to
link up the strata before applying a buffer at a distance of \code{buffer_dist}.
Note that all distances are in the units of \code{sf::st_crs(strata_map)}.
}
\examples{

map <- load_map("bbs_cws")
s <- stratify(by = "bbs_cws", species = "Connecticut Warbler")
p <- prepare_data(s, min_max_route_years = 2)
sp <- prepare_spatial(p, map)

# Visually explore the spatial linkages
sp$map

# Overlay subset strata map on original mapping data
sp <- prepare_spatial(p, map, add_map = map)
sp$map

}
