% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{run_model}
\alias{run_model}
\title{Run JAGS model for prepared species data}
\usage{
run_model(data, init_vals, params, mod, n_chains, n_adapt, n_iter, n_burnin,
  n_thin, parallel)
}
\value{
jagsUI object
}
\description{
\code{run-model} runs a JAGS model as specified by the user for
  the species of interest
}
