% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{run_model}
\alias{run_model}
\title{Run Bayesian model}
\usage{
run_model(
  model_data,
  model,
  model_variant = "hier",
  spatial_neighbours = NULL,
  heavy_tailed = TRUE,
  n_knots = NULL,
  basis = "mgcv",
  use_pois = FALSE,
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_CV = FALSE,
  refresh = 100,
  chains = 4,
  parallel_chains = 4,
  iter_sampling = 1000,
  iter_warmup = 1000,
  adapt_delta = 0.95,
  max_treedepth = 14,
  init_def = NULL,
  out_name = NULL,
  out_dir = ".",
  quiet = FALSE,
  ...
)
}
\arguments{
\item{model_data}{List. Output of \code{prepare_data()}.}

\item{model}{Character.}

\item{...}{}
}
\description{
Run Bayesian model
}
