% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{run_model}
\alias{run_model}
\title{Run Bayesian model}
\usage{
run_model(
  prepped_data,
  model,
  model_variant = "hier",
  spatial_data = NULL,
  heavy_tailed = TRUE,
  n_knots = NULL,
  basis = "mgcv",
  use_pois = FALSE,
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_CV = FALSE,
  refresh = 100,
  chains = 4,
  parallel_chains = 4,
  iter_sampling = 1000,
  iter_warmup = 1000,
  adapt_delta = 0.95,
  max_treedepth = 14,
  init_def = NULL,
  out_name = NULL,
  out_dir = ".",
  save_output = TRUE,
  quiet = FALSE,
  ...
)
}
\arguments{
\item{prepped_data}{List. Output of \code{prepare_data()}.}

\item{model}{Character.}

\item{spatial_data}{List. Output of \code{prepare_spatial()}.}

\item{...}{}
}
\description{
Run Bayesian model
}
\examples{


# You can also specify the GAM model, with an optional number of
# knots to use for the GAM basis.
# By default, the number of knots will be equal to the floor
# of the total unique years for the species / 4
model_data <- prepare_data(strat_data = strat_data,
                           species_to_run = "Pacific Wren",
                           model = "gam",
                           n_knots = 9)

}
