% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{run_model}
\alias{run_model}
\title{Run JAGS model for prepared species data}
\usage{
run_model(jags_data = NULL, model_file_path = NULL, inits = NULL,
  parameters_to_save = c("n"), track_n = TRUE, n_chains = 3,
  n_adapt = 500, n_burnin = 20000, n_thin = 10,
  n_saved_steps = 2000, n_iter = 10000, parallel = FALSE,
  quiet = FALSE, ...)
}
\arguments{
\item{jags_data}{List or environment containing the data to model, as output
by \code{prepare_jags_data}}

\item{model_file_path}{Path to custom model. Overrides the \code{model}
variable set by \code{prepare_jags_data}}

\item{inits}{Optional list of initialization values for JAGS model.
If none are specified, the JAGS model will generate its own
initial values.}

\item{parameters_to_save}{Character vector of parameters to monitor in JAGS. Defaults
to just monitoring "n"}

\item{track_n}{By default, the parameter "n" will always be tracked, even if the user
forgets to specify it. However, if the user is positive they do not want
to track "n", this parameter can be be set to \code{FALSE}. NOTE: you will
not be able to generate annual indices if "n" is not tracked.}

\item{n_chains}{Optional number of chains to run. Defaults to 3.}

\item{n_adapt}{Optional integer specifying the number of steps to
adapt the JAGS model. It is recommended to do a minimum of 100,
\code{run-model} will default to 500.}

\item{n_burnin}{Optional integer specifying the number of iterations
to burn in the model. Defaults to 20000 per chain.}

\item{n_thin}{Optional number of steps to thin or discard.}

\item{n_saved_steps}{Optional number of steps to save per chain.
Defaults to 2000.}

\item{n_iter}{Optional number of iterations per chain. Defaults to 10000.}

\item{parallel}{Should each chain be run parallel on separate cores?
If TRUE, the number of cores used will be the minimum of the
\code{n_chains} specified and the number of cores on your computer}

\item{quiet}{Should JAGS output be suppressed?}

\item{...}{Additional arguments}
}
\value{
jagsUI object
}
\description{
\code{run-model} runs a JAGS model as specified by the user for
  the species of interest
}
\examples{

\dontrun{
# Download BBS data, stratify, and prepare data for a JAGS run with Standard model
bbs_data <- fetch_bbs_data()
data_stratified <- stratify(bbs_data, stratify_by = "latlong")
data_jags <- prepare_jags_data(strat_data = data_stratified,
                               species_to_run = "Bufflehead",
                               model = "standard")

# Run a JAGS model with default JAGS arguments
# Note that you do not have to specify the model file. This is saved
#   in the output from prepare_jags_data
jags_mod <- run_model(jags_data = data_jags)

# You can specify how many chains to run, how many adaptation interations,
# how many burn in iterations, how many sampling interations, and how much\\
# thinning.
jags_mod <- run_model(jags_data = data_jags,
                      n_chains = 2,
                      n_adapt = 200,
                      n_burnin = 5000,
                      n_thin = 10,
                      n_iter = 10000)

# You can specify a vector of variable names to monitor. By default, the
#   variable "n" is monitored to produce trends
jags_mod <- run_model(jags_data = data_jags,
                      parameters_to_save = c("n", "strata"))

# In fact, the variable "n" will always be tracked, even if you don't specify
#  it to be tracked. This ensures that trends can always be calculated in
#  case the user forgets to specify "n" if they went to specify other
#  variables to track. If, however, you are absolutely sure you do not
#  want to track "n", you can set the parameter "track_n" to FALSE

jags_mod <- run_model(jags_data = data_jags,
                      parameters_to_save = ("strata", "beta"),
                      track_n = FALSE)
}

}
