% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{save_model_run}
\alias{save_model_run}
\title{Save output of run_model()}
\usage{
save_model_run(model_output, path = NULL, quiet = FALSE)
}
\arguments{
\item{path}{Character. Optional file path to use for saved data. Defaults to
the file path used for the original run}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}}
}
\description{
This function closely imitate \code{cmdstanr::save_object()} but saves the
entire model output oject from \code{run_model()} which contains more details
regarding stratification etc.
}
