% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-model.R
\name{save_model_run}
\alias{save_model_run}
\title{Save output of run_model()}
\usage{
save_model_run(model_output, path = NULL, quiet = FALSE)
}
\arguments{
\item{model_output}{List. Model output generated by \code{run_model()}.}

\item{path}{Character. Optional file path to use for saved data. Defaults to
the file path used for the original run}

\item{quiet}{Logical. Suppress progress messages? Default \code{FALSE}.}
}
\value{
Nothing. Creates .rds file at \code{path}.
}
\description{
This function closely imitate \code{cmdstanr::save_object()} but saves the
entire model output object from \code{run_model()} which contains more details
regarding stratification etc.
}
\details{
Files are saved to \code{path}, or if not provided to the original location of
the Stan model run files (provided the original files exist).
}
\examples{

# By default, the model is saved as an RDS file during `run_model()`

# But you can also deliberately save the file (here with an example model)
save_model_run(pacific_wren_model, path = "my_model.rds")

# Clean up
unlink("my_model.rds")
}
