% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_sample}
\alias{species_sample}
\title{Sample North American bird species list}
\format{A data frame with 20 rows and 10 variables:
\describe{
  \item{seq}{Sequence - USGS use}
  \item{aou}{Numerical representation of the species, designated by American Ornithological Union}
  \item{english}{Species name in English}
  \item{french}{Species name in French}
  \item{spanish}{Species name in Spanish}
  \item{order}{Taxonomic order}
  \item{family}{Taxonomic family}
  \item{genus}{Taxonomic genus}
  \item{species}{Taxonomic species}
  \item{sp.bbs}{Same as aou, no leading 0}
}}
\usage{
species_sample
}
\description{
A dataset containing species list of North America in 3 languages
  NOTE: This is only a random sample of species data, not the full set.
  The full count set is obtained via the function \code{fetch_bbs_data}.
  The data is obtained from the United States Geological Survey and is
  subject to change as new data is added each year. The data should be cited as follows:
}
\details{
Pardieck, K.L., D.J. Ziolkowski Jr., M. Lutmerding and M.-A.R. Hudson. 2018.
  North American Breeding Bird Survey Dataset 1966 - 2017, version 2017.0. U.S.
  Geological Survey, Patuxent Wildlife Research Center. https://doi.org/10.5066/F76972V8.
}
\keyword{datasets}
