% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify raw Breeding Bird Survey data}
\usage{
stratify(bbs_data, stratify_by = NULL, quiet = FALSE)
}
\arguments{
\item{bbs_data}{Large list of raw BBS data. Can be obtained using
the \code{fetchBBSdata()} function.}

\item{stratify_by}{String argument of stratification type.
Options are "state", "bcr", "latlong", "bbs_cws", "bbs_usgs"}

\item{quiet}{Should progress bars be suppressed?}
}
\value{
Large list (3 elements) consisting of:
  \item{bird_strat}{Dataframe of stratified bird data}
  \item{route_strat}{Dataframe of stratified route data}
  \item{species_strat}{Dataframe of stratified species data}
  \item{stratify_by}{Argument used for stratification}
}
\description{
Assigns each bird count data point and each route a strata
  based on its geographic location and the stratification
  as specified by the user.
}
\examples{

\dontrun{

# Download BBS data and stratify by USGS BBS stratifications.
bbs <- fetch_bbs_data()
data_strat <- stratify(bbs_data = bbs, stratify_by = "bbs_usgs")

# Stratify by Bird Conservation Regions only
# If you don't need the raw BBS data, you can call fetch_bbs_data()
#   directly in the stratify command.
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")

# Stratify by CWS BBS stratifications
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bbs_cws")

# Stratify by State/Province/Territory only
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "state")

# Stratify by blocks of 1 degree of latitude X 1 degree of longitude
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "latlong")
}
}
