% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify raw Breeding Bird Survey data}
\usage{
stratify(bbs_data, stratify_by = "bbs")
}
\arguments{
\item{bbs_data}{Large list of raw BBS data. Can be obtained using
the \code{fetchBBSdata()} function.}

\item{stratify_by}{String argument of stratification type. Defaults
to "bbs", which is the standard BCR X State used in BBS data analysis.
Other options are "state", "bcr", "degree_block"}
}
\value{
Large list (3 elements) consisting of:
  \item{bird_strat}{Dataframe of stratified bird data}
  \item{route_strat}{Dataframe of stratified route data}
  \item{species_strat}{Dataframe of stratified species data}
  \item{stratify_by}{Argument used for stratification}
}
\description{
Assigns each bird count data point and each route a strata
  based on its geographic location and the stratification
  as specified by the user.
}
\examples{

\dontrun{

# Download BBS data and stratify by traditional BBS stratifications.
bbs <- fetch_bbs_data()
data_strat <- stratify(bbs_data = bbs, stratify_by = "bbs")

# Stratify by Bird Conservation Regions only
# If you don't need the raw BBS data, you can call fetch_bbs_data()
#   directly in the stratify command.
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "bcr")

# Stratify by State/Province/Territory only
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_by = "state")

# Stratify by blocks of 1 degree of latitude X 1 degree of longitude
data_strat <- stratify(bbs_data = fetch_bbs_data(), stratify_bs = "degree_block")

# You can choose not to specify a stratification type and default to BBS stratification
data_strat <- stratify(bbs_data = fetch_bbs_data())
}

}
