% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify raw Breeding Bird Survey data}
\usage{
stratify(
  by,
  sample_data = FALSE,
  lump_species_forms = TRUE,
  release = 2022,
  quiet = FALSE
)
}
\arguments{
\item{by}{Character. Stratification type. One of "state", "bcr", "latlong",
"bbs_cws", "bbs_usgs"}

\item{sample_data}{Logical. Use sample data (just Pacific Wren)}

\item{lump_species_forms}{Logical. Whether to lump together species with
multiple forms (default \code{TRUE}). See Details.}

\item{quiet}{Logical. Suppress progress messages}
}
\value{
Large list (3 elements) consisting of:
\item{birds_strat}{Data frame of stratified bird counts data}
\item{routes_strat}{Data frame of stratified routes data}
\item{species_strat}{Data frame of stratified species data}
\item{stratify_by}{Argument used for stratification}
}
\description{
Assigns each bird count data point and each route a strata based on its
geographic location and the stratification as specified by the user.
}
\details{
If \code{lump_species_forms} is \code{TRUE}, species with multiple forms
(e.g., "(unid. race) Dark-eyed Junco"), are duplicated in the data. Once
with individual, identified, forms, and once by lumping all forms together
(including both unidentified and identified forms). This allows users to
specify either individual forms, (e.g., "(Oregon Junco) Dark-eyed Junco")
or grouped forms (e.g., "Dark-eyed Junco (all forms)") in later steps
(e.g., \code{prepare_data()}) without rerunning the \code{stratify()} step. If the
user wishes to keep the unidentified form separate, \code{lump_species_forms}
can be set to \code{FALSE}.
}
\examples{

# Toy examples using Pacific Wren sample data

# Stratify by CWS USGS stratifications
data_strat <- stratify(by = "bbs_usgs", sample_data = TRUE)

# Stratify by Bird Conservation Regions only
data_strat <- stratify(by = "bcr", sample_data = TRUE)

# Stratify by CWS BBS stratifications
data_strat <- stratify(by = "bbs_cws", sample_data = TRUE)

# Stratify by State/Province/Territory only
data_strat <- stratify(by = "state", sample_data = TRUE)

# Stratify by blocks of 1 degree of latitude X 1 degree of longitude
data_strat <- stratify(by = "latlong", sample_data = TRUE)

# To stratify the entire dataset, simply set the sample_data = FALSE,
# or drop that argument altogether. The function then requires fetch_bbs_data()
# to have been run (takes about 10 minutes). Stratification of the entire dataset
# may take up to 3 minutes.

}
