% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratify.R
\name{stratify}
\alias{stratify}
\title{Stratify raw Breeding Bird Survey data}
\usage{
stratify(by = NULL, sample_data = FALSE, bbs_data = NULL,
  quiet = FALSE, stratify_by = NULL)
}
\arguments{
\item{by}{String argument of stratification type.
Options are "state", "bcr", "latlong", "bbs_cws", "bbs_usgs"}

\item{sample_data}{Should just sample data (just Wood Thrush) be used?
Defaults to FALSE}

\item{bbs_data}{Raw BBS data saved as a list of 3 data frames.
Not necessary if you have already run \code{fetch_bbs_data}}

\item{quiet}{Should progress bars be suppressed?}

\item{stratify_by}{Deprecated in favour of 'by'}
}
\value{
Large list (3 elements) consisting of:
  \item{bird_strat}{Dataframe of stratified bird data}
  \item{route_strat}{Dataframe of stratified route data}
  \item{species_strat}{Dataframe of stratified species data}
  \item{by}{Argument used for stratification}
}
\description{
Assigns each bird count data point and each route a strata
  based on its geographic location and the stratification
  as specified by the user.
}
\examples{

\dontrun{

# Download BBS data and stratify by USGS BBS stratifications.
fetch_bbs_data()
data_strat <- stratify(by = "bbs_usgs")

# Stratify by Bird Conservation Regions only
data_strat <- stratify(by = "bcr")

# Stratify by CWS BBS stratifications
data_strat <- stratify(by = "bbs_cws")

# Stratify by State/Province/Territory only
data_strat <- stratify(by = "state")

# Stratify by blocks of 1 degree of latitude X 1 degree of longitude
data_strat <- stratify(by = "latlong")
}
}
