% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waic.R
\name{waic}
\alias{waic}
\title{Calculate Watanabe–Akaike information criterion}
\usage{
waic(jags_data = NULL, jags_mod = NULL)
}
\arguments{
\item{jags_data}{Data prepared by \code{prepare_jags_data}, used
for input to the JAGS model}

\item{jags_mod}{JAGS list generated by \code{run_model}}
}
\value{
Double precision numerical value
}
\description{
\code{waic} returns the Watanabe–Akaike information criterion (WAIC) for the
  supplied model. WAIC is a generalization of the Akaike information
  criterion onto singular statistical model.
}
\details{
NOTE: in order to calculated WAIC, the model MUST track the parameter
  "lambda". In species that are data-rich, such as Barn Swallow,
  this produces extremely large JAGS objects, and takes up a considerable
  amount of memory when simulating with \code{run_model}

See examples for details.
}
\examples{

\dontrun{
  # Run a JAGS model, and be sure to track the parameter "lambda"
  jags_mod <- run_model(jags_data = prepared_data,
                        parameters_to_save = "lambda")

  # Output WAIC
  waic(jags_mod)
}

}
